In progress refactoring of BB2RIFExporter. Includes utility classes, SNF and Beneficiary exporters.
