+++ src/main/java/org/mitre/synthea/export/rif/BB2RIFExporter.java
@@ -1,0 +1,2069 @@
+package org.mitre.synthea.export.rif;
+
+import com.google.common.collect.Table;
+
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.StringWriter;
+import java.math.BigDecimal;
+import java.math.RoundingMode;
+import java.nio.file.Path;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.HashMap;
+import java.util.LinkedHashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.MissingResourceException;
+import java.util.Properties;
+import java.util.TimeZone;
+import java.util.concurrent.atomic.AtomicInteger;
+
+import org.mitre.synthea.export.ExportHelper;
+import org.mitre.synthea.export.Exporter;
+import org.mitre.synthea.export.rif.BB2RIFStructure.CARRIER;
+import org.mitre.synthea.export.rif.BB2RIFStructure.DME;
+import org.mitre.synthea.export.rif.BB2RIFStructure.EXPORT_SUMMARY;
+import org.mitre.synthea.export.rif.BB2RIFStructure.HHA;
+import org.mitre.synthea.export.rif.BB2RIFStructure.HOSPICE;
+import org.mitre.synthea.export.rif.BB2RIFStructure.INPATIENT;
+import org.mitre.synthea.export.rif.BB2RIFStructure.NPI;
+import org.mitre.synthea.export.rif.BB2RIFStructure.OUTPATIENT;
+import org.mitre.synthea.export.rif.BB2RIFStructure.PDE;
+import org.mitre.synthea.export.rif.ExporterBase.ClaimType;
+import org.mitre.synthea.helpers.Config;
+import org.mitre.synthea.helpers.ConsolidatedServicePeriods;
+import org.mitre.synthea.helpers.ConsolidatedServicePeriods.ConsolidatedServicePeriod;
+import org.mitre.synthea.helpers.RandomCollection;
+import org.mitre.synthea.helpers.SimpleCSV;
+import org.mitre.synthea.helpers.Utilities;
+import org.mitre.synthea.world.agents.Clinician;
+import org.mitre.synthea.world.agents.PayerManager;
+import org.mitre.synthea.world.agents.Person;
+import org.mitre.synthea.world.agents.Provider;
+import org.mitre.synthea.world.agents.Provider.ProviderType;
+import org.mitre.synthea.world.concepts.Claim;
+import org.mitre.synthea.world.concepts.Claim.ClaimCost;
+import org.mitre.synthea.world.concepts.Claim.ClaimEntry;
+import org.mitre.synthea.world.concepts.ClinicianSpecialty;
+import org.mitre.synthea.world.concepts.HealthRecord;
+import org.mitre.synthea.world.concepts.HealthRecord.Code;
+import org.mitre.synthea.world.concepts.HealthRecord.Device;
+import org.mitre.synthea.world.concepts.HealthRecord.Encounter;
+import org.mitre.synthea.world.concepts.HealthRecord.EncounterType;
+import org.mitre.synthea.world.concepts.HealthRecord.Entry;
+import org.mitre.synthea.world.concepts.HealthRecord.Medication;
+import org.mitre.synthea.world.concepts.HealthRecord.Supply;
+import org.mitre.synthea.world.geography.CMSStateCodeMapper;
+
+/**
+ * BlueButton 2 RIF format exporter. The data format is described here:
+ * <a href="https://github.com/CMSgov/beneficiary-fhir-data/tree/master/apps/bfd-model">
+ * https://github.com/CMSgov/beneficiary-fhir-data/tree/master/apps/bfd-model</a>.
+ * Workflow:
+ *    This exporter now takes advantage of an analyst-centered, config approach
+ *      to setting exported data.
+ *    Specifically,
+ *    1. An analyst uses a spreadsheet to enter information from official documents
+ *      (see syntax below)
+ *    2. Spreadsheet is exported as a tab-separated values file (TSV)
+ *    3. TSV file is copied to ./src/main/resources/export as bfd_field_values.tsv
+ *        (or whatever is specified in synthea.properties)
+ *    4. when this exporter runs, it will read in the config file
+ *        to set the values needed for the export
+ *    5. the resulting values are then written out to the export files
+ *        in the order the export files require
+ * Syntax for TSV file:
+ *    empty cell   : Not yet analyzed, exporter prints a reminder
+ *    N/A          : Not applicable, ignored by exporter
+ *    (ccccc)      : General comments are enclosed in parenthesis, ignored by exporter
+ *                 : If this is the only content, exporter prints a reminder
+ *                 : If a value precedes the comment, exporter will use value,
+ *                 : and ignore the comment (e.g., 0001 (always 0001 for inpatient) will
+ *                 : result in just 0001)
+ *    vvvvv        : Literal replacement value to be used by exporter (e.g., INSERT)
+ *    a,b,c,d      : A "flat" distribution, exporter will randomly pick one
+ *                 : of the comma delimited values. E.g., B,G (B = brand, G = generic)
+ *                 : exporter randomly chooses either B or G and ignores comment
+ *    Coded        : Implementation in exporter source code, ignored
+ *    [Blank]      : Cell was analyzed and value should be empty
+ */
+public class BB2RIFExporter {
+
+  final RifWriters rifWriters;
+
+  final long claimCutoff;
+  final long snfPDPMCutover;
+
+  final List<LinkedHashMap<String, String>> carrierLookup;
+  final CodeMapper conditionCodeMapper;
+  final CodeMapper medicationCodeMapper;
+  final CodeMapper drgCodeMapper;
+  CodeMapper dmeCodeMapper; // not final so we can mock it for tests
+  final CodeMapper hcpcsCodeMapper;
+  final CodeMapper betosCodeMapper;
+  final CodeMapper snfPPSMapper;
+  final CodeMapper snfPDPMMapper;
+  final CodeMapper snfRevCntrMapper;
+  final CodeMapper hhaRevCntrMapper;
+  final Map<String, RandomCollection<String>> externalCodes;
+  final Map<Integer, Double> pdeOutOfPocketThresholds;
+
+  final CMSStateCodeMapper locationMapper;
+  final StaticFieldConfig staticFieldConfig;
+
+
+  /**
+   * Create the output folder and files. Write headers to each file.
+   */
+  private BB2RIFExporter() {
+    conditionCodeMapper = new CodeMapper("export/condition_code_map.json");
+    medicationCodeMapper = new CodeMapper("export/medication_code_map.json");
+    drgCodeMapper = new CodeMapper("export/drg_code_map.json");
+    dmeCodeMapper = new CodeMapper("export/dme_code_map.json");
+    hcpcsCodeMapper = new CodeMapper("export/hcpcs_code_map.json");
+    betosCodeMapper = new CodeMapper("export/betos_code_map.json");
+    snfPPSMapper = new CodeMapper("export/snf_pps_code_map.json");
+    snfPDPMMapper = new CodeMapper("export/snf_pdpm_code_map.json");
+    snfRevCntrMapper = new CodeMapper("export/snf_rev_cntr_code_map.json");
+    hhaRevCntrMapper = new CodeMapper("export/hha_rev_cntr_code_map.json");
+    locationMapper = new CMSStateCodeMapper();
+    externalCodes = loadExternalCodes();
+    try {
+      String csv = Utilities.readResourceAndStripBOM("payers/carriers.csv");
+      carrierLookup = SimpleCSV.parse(csv);
+      staticFieldConfig = new StaticFieldConfig();
+      rifWriters = prepareOutputFiles();
+      for (String tsvIssue: staticFieldConfig.validateTSV()) {
+        System.out.println(tsvIssue);
+      }
+      SimpleDateFormat format = new SimpleDateFormat("yyyyMMdd");
+      format.setTimeZone(TimeZone.getTimeZone("UTC"));
+      claimCutoff = format.parse(Config.get("exporter.bfd.cutoff_date", "20140529")).getTime();
+      snfPDPMCutover = format.parse("20191001").getTime();
+    } catch (IOException | ParseException e) {
+      // wrap the exception in a runtime exception.
+      // the singleton pattern below doesn't work if the constructor can throw
+      // and if these do throw ioexceptions there's nothing we can do anyway
+      throw new RuntimeException(e);
+    }
+    pdeOutOfPocketThresholds = new HashMap<Integer, Double>();
+    try {
+      String csv = Utilities.readResourceAndStripBOM("costs/pde_oop_thresholds.csv");
+      for (LinkedHashMap<String, String> row : SimpleCSV.parse(csv)) {
+        int year = Integer.parseInt(row.get("YEAR"));
+        double threshold = Double.parseDouble(row.get("THRESHOLD"));
+        pdeOutOfPocketThresholds.put(year, threshold);
+      }
+    } catch (IOException e) {
+      throw new RuntimeException(e);
+    }
+  }
+
+  private static Map<String, RandomCollection<String>> loadExternalCodes() {
+    Map<String, RandomCollection<String>> data = new HashMap<String, RandomCollection<String>>();
+    try {
+      String fileData = Utilities.readResourceAndStripBOM("export/external_codes.csv");
+      List<LinkedHashMap<String, String>> csv = SimpleCSV.parse(fileData);
+      for (LinkedHashMap<String, String> row : csv) {
+        String primary = row.get("primary");
+        String external = row.get("external");
+        double count = Double.parseDouble(row.get("count"));
+        if (!data.containsKey(primary)) {
+          data.put(primary, new RandomCollection<String>());
+        }
+        data.get(primary).add(count, external);
+      }
+    } catch (Exception e) {
+      if (Config.getAsBoolean("exporter.bfd.require_code_maps", true)) {
+        throw new MissingResourceException(
+            "Unable to read external code file 'external_codes.csv'",
+            "BB2RIFExporter", "external_codes.csv");
+      } else {
+        // For testing, the external codes are not present.
+        System.out.println("BB2RIFExporter is running without 'external_codes.csv'");
+      }
+      return null;
+    }
+    return data;
+  }
+
+  <E extends Enum<E>> void setExternalCode(Person person,
+      Map<E,String> fieldValues, E diagnosisCodeKey,
+      E externalCodeKey, E externalVersionKey,
+      E externalPOACodeKey, List<String> presentOnAdmission) {
+    // set the external code...
+    boolean set = setExternalCode(person, fieldValues, diagnosisCodeKey,
+        externalCodeKey, externalVersionKey);
+    // ... and also set the 'present on admission' flag...
+    if (set && externalPOACodeKey != null && presentOnAdmission != null) {
+      String primary = fieldValues.get(diagnosisCodeKey);
+      if (primary != null) {
+        String present = presentOnAdmission.contains(primary) ? "Y" : "U";
+        fieldValues.put(externalPOACodeKey, present);
+      }
+    }
+  }
+
+  <E extends Enum<E>> boolean setExternalCode(Person person,
+      Map<E,String> fieldValues, E diagnosisCodeKey,
+      E externalCodeKey, E externalVersionKey) {
+    String primary = fieldValues.get(diagnosisCodeKey);
+    if (primary != null) {
+      String prefix = primary.substring(0, 3);
+      if (externalCodes != null && externalCodes.containsKey(prefix)) {
+        String externalCode = externalCodes.get(prefix).next(person);
+        fieldValues.put(externalCodeKey, externalCode);
+        fieldValues.put(externalVersionKey, "0");
+        return true;
+      }
+    }
+    return false;
+  }
+
+
+  /**
+   * Create the output folder. Files will be added as needed.
+   */
+  final RifWriters prepareOutputFiles() throws IOException {
+    // Initialize output writers
+    File output = Exporter.getOutputFolder("bfd", null);
+    output.mkdirs();
+    Path outputDirectory = output.toPath();
+
+    return new RifWriters(outputDirectory);
+  }
+
+  private static DateTimeFormatter MANIFEST_TIMESTAMP_FORMAT
+          = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ssX");
+
+  /**
+   * Export a manifest file that lists all of the other BFD files (except the NPI file
+   * which is special).
+   * @throws IOException if something goes wrong
+   */
+  public void exportManifest() throws IOException {
+    File output = Exporter.getOutputFolder("bfd", null);
+    output.mkdirs();
+    StringWriter manifest = new StringWriter();
+    manifest.write("<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"yes\"?>\n");
+    manifest.write("<dataSetManifest xmlns=\"http://cms.hhs.gov/bluebutton/api/schema/ccw-rif/v9\"");
+    manifest.write(String.format(" timestamp=\"%s\" ",
+             java.time.Instant.now()
+                     .atZone(java.time.ZoneId.of("Z"))
+                     .truncatedTo(java.time.temporal.ChronoUnit.SECONDS)
+                     .format(MANIFEST_TIMESTAMP_FORMAT)));
+    manifest.write("sequenceId=\"0\" syntheticData=\"true\">\n");
+    for (Class<?> rifFile: BB2RIFStructure.RIF_FILES) {
+      for (int year: rifWriters.getYears()) {
+        // generics and arrays are weird so need to cast below rather than declare on the array
+        Class<? extends Enum> rifEnum = (Class<? extends Enum>) rifFile;
+        SynchronizedBBLineWriter writer = rifWriters.getWriter(rifEnum, year);
+        if (writer != null) {
+          manifest.write(String.format("  <entry name=\"%s\" type=\"%s\"/>\n",
+                  writer.getFile().getName(), rifEnum.getSimpleName()));
+        }
+      }
+    }
+    manifest.write("</dataSetManifest>");
+    Path manifestPath = output.toPath().resolve("manifest.xml");
+    Exporter.overwriteFile(manifestPath, manifest.toString());
+  }
+
+  /**
+   * Export NPI writer with synthetic providers.
+   * @throws IOException if something goes horribly wrong.
+   */
+  public void exportNPIs() throws IOException {
+    HashMap<NPI, String> fieldValues = new HashMap<>();
+    SynchronizedBBLineWriter rifWriter = rifWriters.getOrCreateWriter(NPI.class,
+            -1, "tsv", "\t");
+
+    for (Provider h : Provider.getProviderList()) {
+
+      // filter - exports only those organizations in use
+      Table<Integer, String, AtomicInteger> utilization = h.getUtilization();
+      int totalEncounters = utilization.column(Provider.ENCOUNTERS).values().stream()
+          .mapToInt(ai -> ai.get()).sum();
+
+      if (totalEncounters > 0) {
+        // export organization
+        fieldValues.clear();
+        fieldValues.put(NPI.NPI, h.npi);
+        fieldValues.put(NPI.ENTITY_TYPE_CODE, "2");
+        fieldValues.put(NPI.EIN, "<UNAVAIL>");
+        fieldValues.put(NPI.ORG_NAME, h.name);
+        rifWriter.writeValues(fieldValues);
+
+        Map<String, ArrayList<Clinician>> clinicians = h.clinicianMap;
+        for (String specialty : clinicians.keySet()) {
+          ArrayList<Clinician> docs = clinicians.get(specialty);
+          for (Clinician doc : docs) {
+            if (doc.getEncounterCount() > 0) {
+              // export each doc
+              Map<String,Object> attributes = doc.getAttributes();
+              fieldValues.clear();
+              fieldValues.put(NPI.NPI, doc.npi);
+              fieldValues.put(NPI.ENTITY_TYPE_CODE, "1");
+              fieldValues.put(NPI.LAST_NAME,
+                  attributes.get(Clinician.LAST_NAME).toString());
+              fieldValues.put(NPI.FIRST_NAME,
+                  attributes.get(Clinician.FIRST_NAME).toString());
+              fieldValues.put(NPI.PREFIX,
+                  attributes.get(Clinician.NAME_PREFIX).toString());
+              fieldValues.put(NPI.CREDENTIALS, "M.D.");
+              rifWriter.writeValues(fieldValues);
+            }
+          }
+        }
+      }
+    }
+  }
+
+  /**
+   * Export the current values of IDs so subsequent runs can use them as a starting point.
+   * @throws IOException if something goes wrong
+   */
+  public void exportEndState() throws IOException {
+    Properties endState = new Properties();
+    endState.setProperty("exporter.bfd.bene_id_start", ExporterBase.nextBeneId.toString());
+    endState.setProperty("exporter.bfd.clm_id_start", ExporterBase.nextClaimId.toString());
+    endState.setProperty("exporter.bfd.clm_grp_id_start", ExporterBase.nextClaimGroupId.toString());
+    endState.setProperty("exporter.bfd.pde_id_start", ExporterBase.nextPdeId.toString());
+    endState.setProperty("exporter.bfd.mbi_start", ExporterBase.nextMbi.toString());
+    endState.setProperty("exporter.bfd.hicn_start", ExporterBase.nextHicn.toString());
+    endState.setProperty("exporter.bfd.fi_doc_cntl_num_start", ExporterBase.nextFiDocCntlNum.toString());
+    endState.setProperty("exporter.bfd.carr_clm_cntl_num_start", ExporterBase.nextCarrClmCntlNum.toString());
+    File outputDir = Exporter.getOutputFolder("bfd", null);
+    FileOutputStream f = new FileOutputStream(new File(outputDir, "end_state.properties"));
+    endState.store(f, "BFD Properties End State");
+    f.close();
+  }
+
+  /**
+   * Export a single person.
+   * @param person the person to export
+   * @param stopTime end time of simulation
+   * @param yearsOfHistory number of years of claims to export
+   * @throws IOException if something goes wrong
+   */
+  public void export(Person person, long stopTime, int yearsOfHistory) throws IOException {
+    Map<EXPORT_SUMMARY, String> exportCounts = new HashMap<>();
+    long startTime = stopTime - Utilities.convertTime("years", yearsOfHistory);
+    BeneficiaryExporter beneExp = new BeneficiaryExporter(startTime, stopTime, this);
+    exportCounts.put(EXPORT_SUMMARY.BENE_ID, beneExp.exportBeneficiary(person));
+    beneExp.exportBeneficiaryHistory(person);
+    exportCounts.put(EXPORT_SUMMARY.INPATIENT_CLAIMS,
+            Long.toString(exportInpatient(person, startTime, stopTime)));
+    exportCounts.put(EXPORT_SUMMARY.OUTPATIENT_CLAIMS,
+            Long.toString(exportOutpatient(person, startTime, stopTime)));
+    exportCounts.put(EXPORT_SUMMARY.CARRIER_CLAIMS,
+            Long.toString(exportCarrier(person, startTime, stopTime)));
+    exportCounts.put(EXPORT_SUMMARY.PDE_CLAIMS,
+            Long.toString(exportPrescription(person, startTime, stopTime)));
+    exportCounts.put(EXPORT_SUMMARY.DME_CLAIMS,
+            Long.toString(exportDME(person, startTime, stopTime)));
+    exportCounts.put(EXPORT_SUMMARY.HHA_CLAIMS,
+            Long.toString(exportHome(person, startTime, stopTime)));
+    exportCounts.put(EXPORT_SUMMARY.HOSPICE_CLAIMS,
+            Long.toString(exportHospice(person, startTime, stopTime)));
+    exportCounts.put(EXPORT_SUMMARY.SNF_CLAIMS,
+            Long.toString(new SNFExporter(startTime, stopTime, this).export(person)));
+    rifWriters.getOrCreateWriter(EXPORT_SUMMARY.class, -1, "csv", ",").writeValues(exportCounts);
+  }
+
+  static String getPartDCostSharingCode(Person person) {
+    double incomeLevel = Double.parseDouble(
+            person.attributes.get(Person.INCOME_LEVEL).toString());
+    if (incomeLevel >= 1.0) {
+      // Beneficiary enrolled in Parts A and/or B, and Part D; no premium or cost sharing subsidy
+      return "09";
+    } else if (incomeLevel >= 0.6) {
+      // Beneficiary enrolled in Parts A and/or B, and Part D; deemed eligible for LIS with 100%
+      // premium subsidy and high copayment
+      return "03";
+    } else if (incomeLevel >= 0.3) {
+      // Beneficiary enrolled in Parts A and/or B, and Part D; deemed eligible for LIS with 100%
+      // premium subsidy and low copayment
+      return "02";
+    }
+    // Beneficiary enrolled in Parts A and/or B, and Part D; deemed eligible for LIS with 100%
+    // premium subsidy and no copayment
+    return "01";
+  }
+
+  // Income level < 0.3
+  private static final RandomCollection<String> incomeBandOneDualCodes = new RandomCollection<>();
+  // 0.3 <= Income level < 0.6
+  private static final RandomCollection<String> incomeBandTwoDualCodes = new RandomCollection<>();
+  // 0.6 <= Income level < 1.0
+  private static final RandomCollection<String> incomeBandThreeDualCodes = new RandomCollection<>();
+
+  static {
+    // Specified Low-Income Medicare Beneficiary (SLMB)-only
+    incomeBandOneDualCodes.add(1.3, "03");
+    // SLMB and full Medicaid coverage, including prescription drugs
+    incomeBandOneDualCodes.add(0.5, "04");
+    // Other dual eligible, but without Medicaid coverage
+    incomeBandOneDualCodes.add(0.007, "09");
+    // Unknown
+    incomeBandOneDualCodes.add(0.05, "99");
+    // Not in code book but present in database
+    incomeBandOneDualCodes.add(0.04, "AA");
+    // Not in code book but present in database
+    incomeBandOneDualCodes.add(0.1, "");
+
+    // QMB and full Medicaid coverage, including prescription drugs
+    incomeBandTwoDualCodes.add(8.2, "02");
+    // Other dual eligible, but without Medicaid coverage
+    incomeBandTwoDualCodes.add(0.007, "09");
+    // Unknown
+    incomeBandTwoDualCodes.add(0.05, "99");
+    // Not in code book but present in database
+    incomeBandTwoDualCodes.add(0.04, "AA");
+    // Not in code book but present in database
+    incomeBandTwoDualCodes.add(0.1, "");
+
+    // Qualified Medicare Beneficiary (QMB)-only
+    incomeBandThreeDualCodes.add(2.2, "01");
+    // Qualifying individuals (QI)
+    incomeBandThreeDualCodes.add(0.8, "06");
+    // Other dual eligible (not QMB, SLMB, QWDI, or QI) with full Medicaid coverage, including
+    // prescription Drugs
+    incomeBandThreeDualCodes.add(2.9, "08");
+    // Other dual eligible, but without Medicaid coverage
+    incomeBandThreeDualCodes.add(0.007, "09");
+    // Unknown
+    incomeBandThreeDualCodes.add(0.05, "99");
+    // Not in code book but present in database
+    incomeBandThreeDualCodes.add(0.04, "AA");
+    // Not in code book but present in database
+    incomeBandThreeDualCodes.add(0.1, "");
+  }
+
+  String getDualEligibilityCode(Person person, int year) {
+    // TBD add support for the following additional code (%-age in brackets is observed
+    // frequency in CMS data):
+    // 00 (15.6%) - Not enrolled in Medicare for the month
+    String partDCostSharingCode = getPartDCostSharingCode(person);
+    if (partDCostSharingCode.equals("03")) {
+      return incomeBandThreeDualCodes.next(person);
+    } else if (partDCostSharingCode.equals("02")) {
+      return incomeBandTwoDualCodes.next(person);
+    } else if (partDCostSharingCode.equals("01")) {
+      return incomeBandOneDualCodes.next(person);
+    } else if (hasESRD(person, year) || isBlind(person)) {
+      return "05"; // (0.001%) Qualified Disabled Working Individual (QDWI)
+    } else {
+      return "NA"; // (68.3%) Non-Medicaid
+    }
+  }
+
+
+  private static final Comparator<Entry> ENTRY_SORTER = new Comparator<Entry>() {
+    @Override
+    public int compare(Entry o1, Entry o2) {
+      return (int)(o2.start - o1.start);
+    }
+  };
+
+  /**
+   * This returns the list of active diagnoses, sorted by most recent first
+   * and oldest last.
+   * @param person patient with the diagnoses.
+   * @return the list of active diagnoses, sorted by most recent first and oldest last.
+   */
+  List<String> getDiagnosesCodes(Person person, long time) {
+    // Collect the active diagnoses at the given time,
+    // keeping only those diagnoses that are mappable.
+    List<Entry> diagnoses = new ArrayList<Entry>();
+    for (Encounter encounter : person.record.encounters) {
+      if (encounter.start <= time) {
+        for (Entry dx : encounter.conditions) {
+          if (dx.start <= time && (dx.stop == 0L || dx.stop > time)) {
+            if (conditionCodeMapper.canMap(dx.codes.get(0).code)) {
+              String mapped = conditionCodeMapper.map(dx.codes.get(0).code, person, true);
+              // Temporarily add the mapped code... we'll remove it later.
+              Code mappedCode = new Code("ICD10", mapped, dx.codes.get(0).display);
+              dx.codes.add(mappedCode);
+              diagnoses.add(dx);
+            }
+          }
+        }
+      }
+    }
+    // Sort them by date and then return only the mapped codes (not the entire Entry).
+    diagnoses.sort(ENTRY_SORTER);
+    List<String> mappedDiagnosisCodes = new ArrayList<String>();
+    for (Entry dx : diagnoses) {
+      mappedDiagnosisCodes.add(dx.codes.remove(dx.codes.size() - 1).code);
+    }
+    return mappedDiagnosisCodes;
+  }
+
+  /**
+   * Determines whether the person has end stage renal disease at the end of the supplied year.
+   * @param person the person
+   * @param year the year
+   * @return true if has ESRD, false otherwise
+   */
+  boolean hasESRD(Person person, int year) {
+    long timestamp = Utilities.convertCalendarYearsToTime(year + 1); // +1 for end of year
+    List<String> mappedDiagnosisCodes = getDiagnosesCodes(person, timestamp);
+    return mappedDiagnosisCodes.contains("N18.6");
+  }
+
+  static boolean isBlind(Person person) {
+    return person.attributes.containsKey(Person.BLINDNESS)
+            && person.attributes.get(Person.BLINDNESS).equals(true);
+  }
+
+
+  /**
+   * Export outpatient claims details for a single person.
+   * @param person the person to export
+   * @param startTime earliest claim date to export
+   * @param stopTime end time of simulation
+   * @return count of claims exported
+   * @throws IOException if something goes wrong
+   */
+  private long exportOutpatient(Person person, long startTime, long stopTime)
+        throws IOException {
+    long claimCount = 0;
+    HashMap<OUTPATIENT, String> fieldValues = new HashMap<>();
+
+    for (Encounter encounter : person.record.encounters) {
+      if (encounter.stop < startTime || encounter.stop < claimCutoff) {
+        continue;
+      }
+      if (!ExporterBase.getClaimTypes(encounter).contains(ClaimType.OUTPATIENT)) {
+        continue;
+      }
+
+      long claimId = ExporterBase.nextClaimId.getAndDecrement();
+      long claimGroupId = ExporterBase.nextClaimGroupId.getAndDecrement();
+      long fiDocId = ExporterBase.nextFiDocCntlNum.getAndDecrement();
+
+      staticFieldConfig.setValues(fieldValues, OUTPATIENT.class, person);
+
+      // The REQUIRED fields
+      fieldValues.put(OUTPATIENT.BENE_ID, (String) person.attributes.get(ExporterBase.BB2_BENE_ID));
+      fieldValues.put(OUTPATIENT.CLM_ID, "" + claimId);
+      fieldValues.put(OUTPATIENT.CLM_GRP_ID, "" + claimGroupId);
+      fieldValues.put(OUTPATIENT.FI_DOC_CLM_CNTL_NUM, "" + fiDocId);
+      fieldValues.put(OUTPATIENT.CLM_FROM_DT, ExporterBase.bb2DateFromTimestamp(encounter.start));
+      fieldValues.put(OUTPATIENT.CLM_THRU_DT, ExporterBase.bb2DateFromTimestamp(encounter.stop));
+      fieldValues.put(OUTPATIENT.NCH_WKLY_PROC_DT,
+              ExporterBase.bb2DateFromTimestamp(ExportHelper.nextFriday(encounter.stop)));
+      fieldValues.put(OUTPATIENT.PRVDR_NUM, encounter.provider.cmsProviderNum);
+      fieldValues.put(OUTPATIENT.AT_PHYSN_NPI, encounter.clinician.npi);
+      fieldValues.put(OUTPATIENT.RNDRNG_PHYSN_NPI, encounter.clinician.npi);
+      fieldValues.put(OUTPATIENT.ORG_NPI_NUM, encounter.provider.npi);
+      fieldValues.put(OUTPATIENT.OP_PHYSN_NPI, encounter.clinician.npi);
+      fieldValues.put(OUTPATIENT.CLM_PMT_AMT, String.format("%.2f",
+              encounter.claim.getTotalClaimCost()));
+      if (encounter.claim.plan == PayerManager.getGovernmentPayer(PayerManager.MEDICARE)
+          .getGovernmentPayerPlan()) {
+        fieldValues.put(OUTPATIENT.NCH_PRMRY_PYR_CLM_PD_AMT, "0");
+      } else {
+        fieldValues.put(OUTPATIENT.NCH_PRMRY_PYR_CLM_PD_AMT,
+                String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+      }
+      fieldValues.put(OUTPATIENT.PRVDR_STATE_CD,
+              locationMapper.getStateCode(encounter.provider.state));
+      // PTNT_DSCHRG_STUS_CD: 1=home, 2=transfer, 3=SNF, 20=died, 30=still here
+      String field = null;
+      if (encounter.ended) {
+        field = "1";
+      } else {
+        field = "30"; // the patient is still here
+      }
+      if (!person.alive(encounter.stop)) {
+        field = "20"; // the patient died before the encounter ended
+      }
+      fieldValues.put(OUTPATIENT.PTNT_DSCHRG_STUS_CD, field);
+      fieldValues.put(OUTPATIENT.CLM_TOT_CHRG_AMT,
+              String.format("%.2f", encounter.claim.getTotalClaimCost()));
+      fieldValues.put(OUTPATIENT.CLM_OP_PRVDR_PMT_AMT,
+              String.format("%.2f", encounter.claim.getTotalClaimCost()));
+      fieldValues.put(OUTPATIENT.REV_CNTR_TOT_CHRG_AMT,
+              String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+      fieldValues.put(OUTPATIENT.REV_CNTR_NCVRD_CHRG_AMT,
+              String.format("%.2f", encounter.claim.getTotalPatientCost()));
+      fieldValues.put(OUTPATIENT.NCH_BENE_PTB_DDCTBL_AMT,
+              String.format("%.2f", encounter.claim.getTotalDeductiblePaid()));
+      fieldValues.put(OUTPATIENT.REV_CNTR_CASH_DDCTBL_AMT,
+              String.format("%.2f", encounter.claim.getTotalDeductiblePaid()));
+      fieldValues.put(OUTPATIENT.REV_CNTR_COINSRNC_WGE_ADJSTD_C,
+              String.format("%.2f", encounter.claim.getTotalCoinsurancePaid()));
+      fieldValues.put(OUTPATIENT.REV_CNTR_PMT_AMT_AMT,
+              String.format("%.2f", encounter.claim.getTotalClaimCost()));
+      fieldValues.put(OUTPATIENT.REV_CNTR_PRVDR_PMT_AMT,
+              String.format("%.2f", encounter.claim.getTotalClaimCost()));
+      fieldValues.put(OUTPATIENT.REV_CNTR_PTNT_RSPNSBLTY_PMT,
+              String.format("%.2f",
+                      encounter.claim.getTotalDeductiblePaid()
+                              .add(encounter.claim.getTotalCoinsurancePaid())));
+      fieldValues.put(OUTPATIENT.REV_CNTR_RDCD_COINSRNC_AMT,
+              String.format("%.2f", encounter.claim.getTotalCoinsurancePaid()));
+
+      String icdReasonCode = null;
+      if (encounter.reason != null) {
+        // If the encounter has a recorded reason, enter the mapped
+        // values into the principle diagnoses code.
+        if (conditionCodeMapper.canMap(encounter.reason.code)) {
+          icdReasonCode = conditionCodeMapper.map(encounter.reason.code, person, true);
+          fieldValues.put(OUTPATIENT.PRNCPAL_DGNS_CD, icdReasonCode);
+        }
+      }
+
+      // Use the active condition diagnoses to enter mapped values
+      // into the diagnoses codes.
+      List<String> mappedDiagnosisCodes = getDiagnosesCodes(person, encounter.stop);
+      boolean noDiagnoses = mappedDiagnosisCodes.isEmpty();
+      if (!noDiagnoses) {
+        int smallest = Math.min(mappedDiagnosisCodes.size(),
+                BB2RIFStructure.outpatientDxFields.length);
+        for (int i = 0; i < smallest; i++) {
+          OUTPATIENT[] dxField = BB2RIFStructure.outpatientDxFields[i];
+          fieldValues.put(dxField[0], mappedDiagnosisCodes.get(i));
+          fieldValues.put(dxField[1], "0"); // 0=ICD10
+        }
+        if (!fieldValues.containsKey(OUTPATIENT.PRNCPAL_DGNS_CD)) {
+          fieldValues.put(OUTPATIENT.PRNCPAL_DGNS_CD, mappedDiagnosisCodes.get(0));
+        }
+      }
+
+      // Check for external code...
+      setExternalCode(person, fieldValues,
+          OUTPATIENT.PRNCPAL_DGNS_CD, OUTPATIENT.ICD_DGNS_E_CD1, OUTPATIENT.ICD_DGNS_E_VRSN_CD1);
+      setExternalCode(person, fieldValues,
+          OUTPATIENT.PRNCPAL_DGNS_CD, OUTPATIENT.FST_DGNS_E_CD, OUTPATIENT.FST_DGNS_E_VRSN_CD);
+
+      // Use the procedures in this encounter to enter mapped values
+      boolean noProcedures = false;
+      if (!encounter.procedures.isEmpty()) {
+        List<HealthRecord.Procedure> mappableProcedures = new ArrayList<>();
+        List<String> mappedProcedureCodes = new ArrayList<>();
+        for (HealthRecord.Procedure procedure : encounter.procedures) {
+          for (HealthRecord.Code code : procedure.codes) {
+            if (conditionCodeMapper.canMap(code.code)) {
+              mappableProcedures.add(procedure);
+              mappedProcedureCodes.add(conditionCodeMapper.map(code.code, person, true));
+              break; // take the first mappable code for each procedure
+            }
+          }
+        }
+        if (!mappableProcedures.isEmpty()) {
+          int smallest = Math.min(mappableProcedures.size(),
+                  BB2RIFStructure.outpatientPxFields.length);
+          for (int i = 0; i < smallest; i++) {
+            OUTPATIENT[] pxField = BB2RIFStructure.outpatientPxFields[i];
+            fieldValues.put(pxField[0], mappedProcedureCodes.get(i));
+            fieldValues.put(pxField[1], "0"); // 0=ICD10
+            fieldValues.put(pxField[2], ExporterBase.bb2DateFromTimestamp(mappableProcedures.get(i).start));
+          }
+        } else {
+          noProcedures = true;
+        }
+      }
+      if (icdReasonCode == null && noDiagnoses && noProcedures) {
+        continue; // skip this encounter
+      }
+      String revCenter = fieldValues.get(OUTPATIENT.REV_CNTR);
+      if (encounter.type.equals(EncounterType.VIRTUAL.toString())) {
+        revCenter = person.randBoolean() ? "0780" : "0789";
+        fieldValues.put(OUTPATIENT.REV_CNTR, revCenter);
+      }
+
+      synchronized (rifWriters.getOrCreateWriter(OUTPATIENT.class)) {
+        int claimLine = 1;
+        for (ClaimEntry lineItem : encounter.claim.items) {
+          String hcpcsCode = null;
+          if (lineItem.entry instanceof HealthRecord.Procedure) {
+            for (HealthRecord.Code code : lineItem.entry.codes) {
+              if (hcpcsCodeMapper.canMap(code.code)) {
+                hcpcsCode = hcpcsCodeMapper.map(code.code, person, true);
+                break; // take the first mappable code for each procedure
+              }
+            }
+            fieldValues.put(OUTPATIENT.REV_CNTR, revCenter);
+            fieldValues.remove(OUTPATIENT.REV_CNTR_IDE_NDC_UPC_NUM);
+            fieldValues.remove(OUTPATIENT.REV_CNTR_NDC_QTY);
+            fieldValues.remove(OUTPATIENT.REV_CNTR_NDC_QTY_QLFR_CD);
+          } else if (lineItem.entry instanceof HealthRecord.Medication) {
+            HealthRecord.Medication med = (HealthRecord.Medication) lineItem.entry;
+            if (med.administration) {
+              hcpcsCode = "T1502";  // Administration of medication
+              fieldValues.put(OUTPATIENT.REV_CNTR, "0636"); // Drugs requiring specific id
+              String ndcCode = medicationCodeMapper.map(med.codes.get(0).code, person);
+              fieldValues.put(OUTPATIENT.REV_CNTR_IDE_NDC_UPC_NUM, ndcCode);
+              fieldValues.put(OUTPATIENT.REV_CNTR_NDC_QTY, "1"); // 1 Unit
+              fieldValues.put(OUTPATIENT.REV_CNTR_NDC_QTY_QLFR_CD, "UN"); // Unit
+            }
+          }
+          if (hcpcsCode == null) {
+            continue;
+          }
+
+          fieldValues.put(OUTPATIENT.CLM_LINE_NUM, Integer.toString(claimLine++));
+          fieldValues.put(OUTPATIENT.REV_CNTR_DT, ExporterBase.bb2DateFromTimestamp(lineItem.entry.start));
+          fieldValues.put(OUTPATIENT.HCPCS_CD, hcpcsCode);
+          fieldValues.put(OUTPATIENT.REV_CNTR_RATE_AMT,
+              String.format("%.2f", (lineItem.cost)));
+          fieldValues.put(OUTPATIENT.REV_CNTR_PMT_AMT_AMT,
+              String.format("%.2f", lineItem.coinsurancePaidByPayer.add(lineItem.paidByPayer)));
+          fieldValues.put(OUTPATIENT.REV_CNTR_TOT_CHRG_AMT,
+              String.format("%.2f", lineItem.cost));
+          fieldValues.put(OUTPATIENT.REV_CNTR_NCVRD_CHRG_AMT,
+              String.format("%.2f", lineItem.copayPaidByPatient
+              .add(lineItem.deductiblePaidByPatient).add(lineItem.patientOutOfPocket)));
+          rifWriters.writeValues(OUTPATIENT.class, fieldValues);
+        }
+
+        if (claimLine == 1) {
+          // If claimLine still equals 1, then no line items were successfully added.
+          // Add a single top-level entry.
+          fieldValues.put(OUTPATIENT.CLM_LINE_NUM, Integer.toString(claimLine));
+          fieldValues.put(OUTPATIENT.REV_CNTR_DT, ExporterBase.bb2DateFromTimestamp(encounter.start));
+          // 99241: "Office consultation for a new or established patient"
+          fieldValues.put(OUTPATIENT.HCPCS_CD, "99241");
+          rifWriters.writeValues(OUTPATIENT.class, fieldValues);
+        }
+      }
+      claimCount++;
+    }
+    return claimCount;
+  }
+
+  /**
+   * Export inpatient claims details for a single person.
+   * @param person the person to export
+   * @param startTime earliest claim date to export
+   * @param stopTime end time of simulation
+   * @return count of claims exported
+   * @throws IOException if something goes wrong
+   */
+  private long exportInpatient(Person person, long startTime, long stopTime)
+        throws IOException {
+    HashMap<INPATIENT, String> fieldValues = new HashMap<>();
+    long claimCount = 0;
+    boolean previousEmergency = false;
+
+    for (Encounter encounter : person.record.encounters) {
+      if (encounter.stop < startTime || encounter.stop < claimCutoff) {
+        continue;
+      }
+      if (!ExporterBase.getClaimTypes(encounter).contains(ClaimType.INPATIENT)) {
+        previousEmergency = false;
+        continue;
+      }
+
+      long claimId = ExporterBase.nextClaimId.getAndDecrement();
+      long claimGroupId = ExporterBase.nextClaimGroupId.getAndDecrement();
+      long fiDocId = ExporterBase.nextFiDocCntlNum.getAndDecrement();
+
+      fieldValues.clear();
+      staticFieldConfig.setValues(fieldValues, INPATIENT.class, person);
+
+      // The REQUIRED fields
+      fieldValues.put(INPATIENT.BENE_ID, (String) person.attributes.get(ExporterBase.BB2_BENE_ID));
+      fieldValues.put(INPATIENT.CLM_ID, "" + claimId);
+      fieldValues.put(INPATIENT.CLM_GRP_ID, "" + claimGroupId);
+      fieldValues.put(INPATIENT.FI_DOC_CLM_CNTL_NUM, "" + fiDocId);
+      fieldValues.put(INPATIENT.CLM_FROM_DT, ExporterBase.bb2DateFromTimestamp(encounter.start));
+      fieldValues.put(INPATIENT.CLM_ADMSN_DT, ExporterBase.bb2DateFromTimestamp(encounter.start));
+      fieldValues.put(INPATIENT.CLM_THRU_DT, ExporterBase.bb2DateFromTimestamp(encounter.stop));
+      fieldValues.put(INPATIENT.NCH_BENE_DSCHRG_DT, ExporterBase.bb2DateFromTimestamp(encounter.stop));
+      fieldValues.put(INPATIENT.NCH_WKLY_PROC_DT,
+              ExporterBase.bb2DateFromTimestamp(ExportHelper.nextFriday(encounter.stop)));
+      fieldValues.put(INPATIENT.PRVDR_NUM, encounter.provider.cmsProviderNum);
+      fieldValues.put(INPATIENT.AT_PHYSN_NPI, encounter.clinician.npi);
+      fieldValues.put(INPATIENT.ORG_NPI_NUM, encounter.provider.npi);
+      fieldValues.put(INPATIENT.OP_PHYSN_NPI, encounter.clinician.npi);
+      fieldValues.put(INPATIENT.CLM_PMT_AMT,
+              String.format("%.2f", encounter.claim.getTotalClaimCost()));
+      if (encounter.claim.plan == PayerManager.getGovernmentPayer(PayerManager.MEDICARE)
+          .getGovernmentPayerPlan()) {
+        fieldValues.put(INPATIENT.NCH_PRMRY_PYR_CLM_PD_AMT, "0");
+      } else {
+        fieldValues.put(INPATIENT.NCH_PRMRY_PYR_CLM_PD_AMT,
+                String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+      }
+      fieldValues.put(INPATIENT.PRVDR_STATE_CD,
+              locationMapper.getStateCode(encounter.provider.state));
+      // PTNT_DSCHRG_STUS_CD: 1=home, 2=transfer, 3=SNF, 20=died, 30=still here
+      String field = null;
+      String patientStatus = null;
+      if (encounter.ended) {
+        field = "1"; // TODO 2=transfer if the next encounter is also inpatient
+        patientStatus = "A"; // discharged
+      } else {
+        field = "30"; // the patient is still here
+        patientStatus = "C"; // still a patient
+      }
+      if (!person.alive(encounter.stop)) {
+        field = "20"; // the patient died before the encounter ended
+        patientStatus = "B"; // died
+      }
+      fieldValues.put(INPATIENT.PTNT_DSCHRG_STUS_CD, field);
+      fieldValues.put(INPATIENT.NCH_PTNT_STATUS_IND_CD, patientStatus);
+      fieldValues.put(INPATIENT.CLM_TOT_CHRG_AMT,
+              String.format("%.2f", encounter.claim.getTotalClaimCost()));
+      boolean isEmergency = encounter.type.equals(EncounterType.EMERGENCY.toString());
+      if (isEmergency) {
+        field = "1"; // emergency
+        fieldValues.put(INPATIENT.REV_CNTR, "0450"); // emergency
+      } else if (previousEmergency) {
+        field = "2"; // urgent
+      } else {
+        field = "3"; // elective
+      }
+      fieldValues.put(INPATIENT.CLM_IP_ADMSN_TYPE_CD, field);
+      fieldValues.put(INPATIENT.NCH_BENE_IP_DDCTBL_AMT,
+          String.format("%.2f", encounter.claim.getTotalDeductiblePaid()));
+      fieldValues.put(INPATIENT.NCH_BENE_PTA_COINSRNC_LBLTY_AM,
+          String.format("%.2f", encounter.claim.getTotalCoinsurancePaid()));
+      fieldValues.put(INPATIENT.NCH_IP_NCVRD_CHRG_AMT,
+          String.format("%.2f", encounter.claim.getTotalPatientCost()));
+      fieldValues.put(INPATIENT.NCH_IP_TOT_DDCTN_AMT,
+          String.format("%.2f", encounter.claim.getTotalPatientCost()));
+      int days = (int) ((encounter.stop - encounter.start) / (1000 * 60 * 60 * 24));
+      fieldValues.put(INPATIENT.CLM_UTLZTN_DAY_CNT, "" + days);
+      if (days > 60) {
+        field = "" + (days - 60);
+      } else {
+        field = "0";
+      }
+      fieldValues.put(INPATIENT.BENE_TOT_COINSRNC_DAYS_CNT, field);
+      if (days > 60) {
+        field = "1"; // days outlier
+      } else if (encounter.claim.getTotalClaimCost().compareTo(BigDecimal.valueOf(100_000)) > 0) {
+        field = "2"; // cost outlier
+      } else {
+        field = "0"; // no outlier
+      }
+      fieldValues.put(INPATIENT.CLM_DRG_OUTLIER_STAY_CD, field);
+      fieldValues.put(INPATIENT.REV_CNTR_TOT_CHRG_AMT,
+          String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+      fieldValues.put(INPATIENT.REV_CNTR_NCVRD_CHRG_AMT,
+          String.format("%.2f", encounter.claim.getTotalPatientCost()));
+
+      // OPTIONAL FIELDS
+      fieldValues.put(INPATIENT.RNDRNG_PHYSN_NPI, encounter.clinician.npi);
+
+      String icdReasonCode = null;
+      if (encounter.reason != null) {
+        // If the encounter has a recorded reason, enter the mapped
+        // values into the principle diagnoses code.
+        if (conditionCodeMapper.canMap(encounter.reason.code)) {
+          icdReasonCode = conditionCodeMapper.map(encounter.reason.code, person, true);
+          fieldValues.put(INPATIENT.PRNCPAL_DGNS_CD, icdReasonCode);
+          fieldValues.put(INPATIENT.ADMTG_DGNS_CD, icdReasonCode);
+        }
+      }
+      // Use the active condition diagnoses to enter mapped values
+      // into the diagnoses codes.
+      List<String> presentOnAdmission = getDiagnosesCodes(person, encounter.start);
+      List<String> mappedDiagnosisCodes = getDiagnosesCodes(person, encounter.stop);
+      boolean noDiagnoses = mappedDiagnosisCodes.isEmpty();
+      if (!noDiagnoses) {
+        int smallest = Math.min(mappedDiagnosisCodes.size(),
+                BB2RIFStructure.inpatientDxFields.length);
+        for (int i = 0; i < smallest; i++) {
+          INPATIENT[] dxField = BB2RIFStructure.inpatientDxFields[i];
+          String dxCode = mappedDiagnosisCodes.get(i);
+          fieldValues.put(dxField[0], dxCode);
+          fieldValues.put(dxField[1], "0"); // 0=ICD10
+          String present = presentOnAdmission.contains(dxCode) ? "Y" : "N";
+          fieldValues.put(dxField[2], present);
+        }
+        if (!fieldValues.containsKey(INPATIENT.PRNCPAL_DGNS_CD)) {
+          fieldValues.put(INPATIENT.PRNCPAL_DGNS_CD, mappedDiagnosisCodes.get(0));
+        }
+      }
+
+      if (fieldValues.containsKey(INPATIENT.PRNCPAL_DGNS_CD)) {
+        String icdCode = fieldValues.get(INPATIENT.PRNCPAL_DGNS_CD);
+        // Add a DRG code, if applicable
+        if (drgCodeMapper.canMap(icdCode)) {
+          fieldValues.put(INPATIENT.CLM_DRG_CD, drgCodeMapper.map(icdCode, person));
+        }
+        // Check for external code...
+        setExternalCode(person, fieldValues,
+            INPATIENT.PRNCPAL_DGNS_CD, INPATIENT.ICD_DGNS_E_CD1, INPATIENT.ICD_DGNS_E_VRSN_CD1,
+            INPATIENT.CLM_E_POA_IND_SW1, presentOnAdmission);
+        setExternalCode(person, fieldValues,
+            INPATIENT.PRNCPAL_DGNS_CD, INPATIENT.FST_DGNS_E_CD, INPATIENT.FST_DGNS_E_VRSN_CD);
+      }
+
+      // Use the procedures in this encounter to enter mapped values
+      boolean noProcedures = false;
+      if (!encounter.procedures.isEmpty()) {
+        List<HealthRecord.Procedure> mappableProcedures = new ArrayList<>();
+        List<String> mappedProcedureCodes = new ArrayList<>();
+        for (HealthRecord.Procedure procedure : encounter.procedures) {
+          for (HealthRecord.Code code : procedure.codes) {
+            if (conditionCodeMapper.canMap(code.code)) {
+              mappableProcedures.add(procedure);
+              mappedProcedureCodes.add(conditionCodeMapper.map(code.code, person, true));
+              break; // take the first mappable code for each procedure
+            }
+          }
+        }
+        if (!mappableProcedures.isEmpty()) {
+          int smallest = Math.min(mappableProcedures.size(),
+                  BB2RIFStructure.inpatientPxFields.length);
+          for (int i = 0; i < smallest; i++) {
+            INPATIENT[] pxField = BB2RIFStructure.inpatientPxFields[i];
+            fieldValues.put(pxField[0], mappedProcedureCodes.get(i));
+            fieldValues.put(pxField[1], "0"); // 0=ICD10
+            fieldValues.put(pxField[2], ExporterBase.bb2DateFromTimestamp(mappableProcedures.get(i).start));
+          }
+        } else {
+          noProcedures = true;
+        }
+      }
+      if (icdReasonCode == null && noDiagnoses && noProcedures) {
+        continue; // skip this encounter
+      }
+      previousEmergency = isEmergency;
+      String revCenter = fieldValues.get(INPATIENT.REV_CNTR);
+
+      synchronized (rifWriters.getOrCreateWriter(INPATIENT.class)) {
+        int claimLine = 1;
+        for (ClaimEntry lineItem : encounter.claim.items) {
+          String hcpcsCode = null;
+          if (lineItem.entry instanceof HealthRecord.Procedure) {
+            for (HealthRecord.Code code : lineItem.entry.codes) {
+              if (hcpcsCodeMapper.canMap(code.code)) {
+                hcpcsCode = hcpcsCodeMapper.map(code.code, person, true);
+                break; // take the first mappable code for each procedure
+              }
+            }
+            fieldValues.put(INPATIENT.REV_CNTR, revCenter);
+            fieldValues.remove(INPATIENT.REV_CNTR_NDC_QTY);
+            fieldValues.remove(INPATIENT.REV_CNTR_NDC_QTY_QLFR_CD);
+          } else if (lineItem.entry instanceof HealthRecord.Medication) {
+            HealthRecord.Medication med = (HealthRecord.Medication) lineItem.entry;
+            if (med.administration) {
+              hcpcsCode = "T1502";  // Administration of medication
+              fieldValues.put(INPATIENT.REV_CNTR, "0250"); // Pharmacy-general classification
+              fieldValues.put(INPATIENT.REV_CNTR_NDC_QTY, "1"); // 1 Unit
+              fieldValues.put(INPATIENT.REV_CNTR_NDC_QTY_QLFR_CD, "UN"); // Unit
+            }
+          }
+          if (hcpcsCode == null) {
+            continue;
+          }
+
+          fieldValues.put(INPATIENT.CLM_LINE_NUM, Integer.toString(claimLine++));
+          fieldValues.put(INPATIENT.HCPCS_CD, hcpcsCode);
+          fieldValues.put(INPATIENT.REV_CNTR_UNIT_CNT, "" + Integer.max(1, days));
+          BigDecimal rate = lineItem.cost.divide(
+                  BigDecimal.valueOf(Integer.max(1, days)), RoundingMode.HALF_EVEN)
+                  .setScale(2, RoundingMode.HALF_EVEN);
+          fieldValues.put(INPATIENT.REV_CNTR_RATE_AMT,
+              String.format("%.2f", rate));
+          fieldValues.put(INPATIENT.REV_CNTR_TOT_CHRG_AMT,
+              String.format("%.2f", lineItem.cost));
+          fieldValues.put(INPATIENT.REV_CNTR_NCVRD_CHRG_AMT,
+              String.format("%.2f", lineItem.copayPaidByPatient
+              .add(lineItem.deductiblePaidByPatient).add(lineItem.patientOutOfPocket)));
+          if (lineItem.patientOutOfPocket.compareTo(Claim.ZERO_CENTS) == 0
+                  && lineItem.deductiblePaidByPatient.compareTo(Claim.ZERO_CENTS) == 0) {
+            // Not subject to deductible or coinsurance
+            fieldValues.put(INPATIENT.REV_CNTR_DDCTBL_COINSRNC_CD, "3");
+          } else if (lineItem.patientOutOfPocket.compareTo(Claim.ZERO_CENTS) > 0
+                  && lineItem.deductiblePaidByPatient.compareTo(Claim.ZERO_CENTS) > 0) {
+            // Subject to deductible and coinsurance
+            fieldValues.put(INPATIENT.REV_CNTR_DDCTBL_COINSRNC_CD, "0");
+          } else if (lineItem.patientOutOfPocket.compareTo(Claim.ZERO_CENTS) == 0) {
+            // Not subject to deductible
+            fieldValues.put(INPATIENT.REV_CNTR_DDCTBL_COINSRNC_CD, "1");
+          } else {
+            // Not subject to coinsurance
+            fieldValues.put(INPATIENT.REV_CNTR_DDCTBL_COINSRNC_CD, "2");
+          }
+          rifWriters.writeValues(INPATIENT.class, fieldValues);
+        }
+
+        if (claimLine == 1) {
+          // If claimLine still equals 1, then no line items were successfully added.
+          // Add a single top-level entry.
+          fieldValues.put(INPATIENT.CLM_LINE_NUM, Integer.toString(claimLine));
+          // HCPCS 99221: "Inpatient hospital visits: Initial and subsequent"
+          fieldValues.put(INPATIENT.HCPCS_CD, "99221");
+          rifWriters.writeValues(INPATIENT.class, fieldValues);
+        }
+      }
+      claimCount++;
+    }
+    return claimCount;
+  }
+
+  /**
+   * Export carrier claims details for a single person.
+   * @param person the person to export
+   * @param startTime earliest claim date to export
+   * @param stopTime end time of simulation
+   * @return number of claims exported
+   * @throws IOException if something goes wrong
+   */
+  private long exportCarrier(Person person, long startTime, long stopTime) throws IOException {
+    HashMap<CARRIER, String> fieldValues = new HashMap<>();
+
+    long claimCount = 0;
+    double latestHemoglobin = 0;
+
+    for (Encounter encounter : person.record.encounters) {
+      if (encounter.stop < startTime || encounter.stop < claimCutoff) {
+        continue;
+      }
+      if (!ExporterBase.getClaimTypes(encounter).contains(ClaimType.CARRIER)) {
+        continue;
+      }
+
+      long claimId = ExporterBase.nextClaimId.getAndDecrement();
+      long claimGroupId = ExporterBase.nextClaimGroupId.getAndDecrement();
+      long carrClmId = ExporterBase.nextCarrClmCntlNum.getAndDecrement();
+
+      for (HealthRecord.Observation observation : encounter.observations) {
+        if (observation.containsCode("718-7", "http://loinc.org")) {
+          latestHemoglobin = (double) observation.value;
+        }
+      }
+
+      staticFieldConfig.setValues(fieldValues, CARRIER.class, person);
+      fieldValues.put(CARRIER.BENE_ID, (String) person.attributes.get(ExporterBase.BB2_BENE_ID));
+
+      // The REQUIRED fields
+      fieldValues.put(CARRIER.CLM_ID, "" + claimId);
+      fieldValues.put(CARRIER.CLM_GRP_ID, "" + claimGroupId);
+      fieldValues.put(CARRIER.CARR_CLM_CNTL_NUM, "" + carrClmId);
+      fieldValues.put(CARRIER.CLM_FROM_DT, ExporterBase.bb2DateFromTimestamp(encounter.start));
+      fieldValues.put(CARRIER.LINE_1ST_EXPNS_DT, ExporterBase.bb2DateFromTimestamp(encounter.start));
+      fieldValues.put(CARRIER.CLM_THRU_DT, ExporterBase.bb2DateFromTimestamp(encounter.stop));
+      fieldValues.put(CARRIER.LINE_LAST_EXPNS_DT, ExporterBase.bb2DateFromTimestamp(encounter.stop));
+      fieldValues.put(CARRIER.NCH_WKLY_PROC_DT,
+              ExporterBase.bb2DateFromTimestamp(ExportHelper.nextFriday(encounter.stop)));
+      fieldValues.put(CARRIER.CARR_NUM,
+              getCarrier(encounter.provider.state, CARRIER.CARR_NUM));
+      fieldValues.put(CARRIER.CLM_PMT_AMT,
+              String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+      if (encounter.claim.plan == PayerManager.getGovernmentPayer(PayerManager.MEDICARE)
+          .getGovernmentPayerPlan()) {
+        fieldValues.put(CARRIER.CARR_CLM_PRMRY_PYR_PD_AMT, "0");
+      } else {
+        fieldValues.put(CARRIER.CARR_CLM_PRMRY_PYR_PD_AMT,
+                String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+      }
+      // NCH_CLM_BENE_PMT_AMT, is always zero (set in field value spreadsheet)
+      fieldValues.put(CARRIER.NCH_CLM_PRVDR_PMT_AMT,
+              String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+      fieldValues.put(CARRIER.NCH_CARR_CLM_SBMTD_CHRG_AMT,
+              String.format("%.2f", encounter.claim.getTotalClaimCost()));
+      fieldValues.put(CARRIER.NCH_CARR_CLM_ALOWD_AMT,
+              String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+      fieldValues.put(CARRIER.CARR_CLM_CASH_DDCTBL_APLD_AMT,
+              String.format("%.2f", encounter.claim.getTotalDeductiblePaid()));
+      fieldValues.put(CARRIER.CARR_CLM_RFRNG_PIN_NUM, encounter.provider.cmsPin);
+      fieldValues.put(CARRIER.CARR_PRFRNG_PIN_NUM, encounter.provider.cmsPin);
+      fieldValues.put(CARRIER.ORG_NPI_NUM, encounter.provider.npi);
+      fieldValues.put(CARRIER.PRF_PHYSN_NPI, encounter.clinician.npi);
+      fieldValues.put(CARRIER.RFR_PHYSN_NPI, encounter.clinician.npi);
+      fieldValues.put(CARRIER.PRVDR_SPCLTY,
+          ClinicianSpecialty.getCMSProviderSpecialtyCode(
+              (String) encounter.clinician.attributes.get(Clinician.SPECIALTY)));
+      fieldValues.put(CARRIER.TAX_NUM,
+              BeneficiaryExporter.bb2TaxId((String)encounter.clinician.attributes.get(Person.IDENTIFIER_SSN)));
+      fieldValues.put(CARRIER.LINE_SRVC_CNT, "" + encounter.claim.items.size());
+      fieldValues.put(CARRIER.CARR_LINE_PRCNG_LCLTY_CD,
+              getCarrier(encounter.provider.state, CARRIER.CARR_LINE_PRCNG_LCLTY_CD));
+      // length of encounter in minutes
+      fieldValues.put(CARRIER.CARR_LINE_MTUS_CNT,
+              "" + ((encounter.stop - encounter.start) / (1000 * 60)));
+
+      fieldValues.put(CARRIER.LINE_HCT_HGB_RSLT_NUM,
+              "" + latestHemoglobin);
+      fieldValues.put(CARRIER.LINE_PLACE_OF_SRVC_CD, getPlaceOfService(encounter));
+
+      // OPTIONAL
+      fieldValues.put(CARRIER.PRF_PHYSN_UPIN, encounter.provider.cmsUpin);
+      fieldValues.put(CARRIER.RFR_PHYSN_UPIN, encounter.provider.cmsUpin);
+      fieldValues.put(CARRIER.PRVDR_STATE_CD, encounter.provider.state);
+      fieldValues.put(CARRIER.PRVDR_ZIP, encounter.provider.zip);
+
+      String icdReasonCode = null;
+      if (encounter.reason != null) {
+        // If the encounter has a recorded reason, enter the mapped
+        // values into the principle diagnoses code.
+        if (conditionCodeMapper.canMap(encounter.reason.code)) {
+          icdReasonCode = conditionCodeMapper.map(encounter.reason.code, person, true);
+          fieldValues.put(CARRIER.PRNCPAL_DGNS_CD, icdReasonCode);
+          fieldValues.put(CARRIER.LINE_ICD_DGNS_CD, icdReasonCode);
+        }
+      }
+
+      // Use the active condition diagnoses to enter mapped values
+      // into the diagnoses codes.
+      List<String> mappedDiagnosisCodes = getDiagnosesCodes(person, encounter.stop);
+      if (mappedDiagnosisCodes.isEmpty() && icdReasonCode == null) {
+        continue; // skip this encounter
+      }
+      int smallest = Math.min(mappedDiagnosisCodes.size(),
+              BB2RIFStructure.carrierDxFields.length);
+      for (int i = 0; i < smallest; i++) {
+        CARRIER[] dxField = BB2RIFStructure.carrierDxFields[i];
+        fieldValues.put(dxField[0], mappedDiagnosisCodes.get(i));
+        fieldValues.put(dxField[1], "0"); // 0=ICD10
+      }
+      if (!fieldValues.containsKey(CARRIER.PRNCPAL_DGNS_CD)) {
+        fieldValues.put(CARRIER.PRNCPAL_DGNS_CD, mappedDiagnosisCodes.get(0));
+      }
+
+      synchronized (rifWriters.getOrCreateWriter(CARRIER.class)) {
+        int lineNum = 1;
+        CLIA cliaLab = ExporterBase.cliaLabNumbers[person.randInt(ExporterBase.cliaLabNumbers.length)];
+        List<ClaimEntry> allItems = new ArrayList<>();
+        allItems.add(encounter.claim.mainEntry);
+        allItems.addAll(encounter.claim.items);
+        for (ClaimEntry lineItem : allItems) {
+          String hcpcsCode = "";
+          String ndcCode = "";
+          if (lineItem.entry instanceof HealthRecord.Procedure) {
+            for (HealthRecord.Code code : lineItem.entry.codes) {
+              if (hcpcsCodeMapper.canMap(code.code)) {
+                hcpcsCode = hcpcsCodeMapper.map(code.code, person, true);
+                break; // take the first mappable code for each procedure
+              }
+            }
+          } else if (lineItem.entry instanceof HealthRecord.Medication) {
+            HealthRecord.Medication med = (HealthRecord.Medication) lineItem.entry;
+            if (med.administration) {
+              hcpcsCode = "T1502";  // Administration of medication
+              ndcCode = medicationCodeMapper.map(med.codes.get(0).code, person);
+            }
+          }
+          if (icdReasonCode == null) {
+            // If there is an icdReasonCode, then then LINE_ICD_DGNS_CD is already set.
+            // If not, we might choose a value for each line item.
+            double probability = person.rand();
+            if (probability <= 0.06) {
+              // Random code
+              int index = person.randInt(mappedDiagnosisCodes.size());
+              String code = mappedDiagnosisCodes.get(index);
+              fieldValues.put(CARRIER.LINE_ICD_DGNS_CD, code);
+            } else if (probability <= 0.48) {
+              // The principal diagnosis code
+              fieldValues.put(CARRIER.LINE_ICD_DGNS_CD, fieldValues.get(CARRIER.PRNCPAL_DGNS_CD));
+            } else {
+              // No line item diagnosis code
+              fieldValues.remove(CARRIER.LINE_ICD_DGNS_CD);
+            }
+          }
+          // TBD: decide whether line item skip logic is needed here and in other files
+          // TBD: affects ~80% of carrier claim lines, so left out for now
+          // if (hcpcsCode == null) {
+          //   continue; // skip this line item
+          // }
+          fieldValues.put(CARRIER.HCPCS_CD, hcpcsCode);
+          if (betosCodeMapper.canMap(hcpcsCode)) {
+            fieldValues.put(CARRIER.BETOS_CD, betosCodeMapper.map(hcpcsCode, person));
+          } else {
+            fieldValues.put(CARRIER.BETOS_CD, "");
+          }
+          fieldValues.put(CARRIER.LINE_NDC_CD, ndcCode);
+          fieldValues.put(CARRIER.LINE_BENE_PTB_DDCTBL_AMT,
+                  String.format("%.2f", lineItem.deductiblePaidByPatient));
+          fieldValues.put(CARRIER.LINE_COINSRNC_AMT,
+                  String.format("%.2f", lineItem.coinsurancePaidByPayer));
+
+          // Like NCH_CLM_BENE_PMT_AMT, LINE_BENE_PMT_AMT is always zero
+          // (set in field value spreadsheet)
+          BigDecimal providerAmount = lineItem.coinsurancePaidByPayer.add(lineItem.paidByPayer);
+          fieldValues.put(CARRIER.LINE_PRVDR_PMT_AMT,
+              String.format("%.2f", providerAmount));
+          fieldValues.put(CARRIER.LINE_NCH_PMT_AMT,
+              String.format("%.2f", providerAmount));
+          fieldValues.put(CARRIER.LINE_SBMTD_CHRG_AMT,
+              String.format("%.2f", lineItem.cost));
+          fieldValues.put(CARRIER.LINE_ALOWD_CHRG_AMT,
+              String.format("%.2f", lineItem.cost.subtract(lineItem.adjustment)));
+
+          // If this item is a lab report, add the number of the clinical lab...
+          if  (lineItem.entry instanceof HealthRecord.Report) {
+            if (encounter.provider.cliaNumber != null) {
+              fieldValues.put(CARRIER.CARR_LINE_CLIA_LAB_NUM, encounter.provider.cliaNumber);
+            } else {
+              fieldValues.put(CARRIER.CARR_LINE_CLIA_LAB_NUM, cliaLab.toString());
+            }
+          }
+
+          // set the line number and write out field values
+          fieldValues.put(CARRIER.LINE_NUM, Integer.toString(lineNum++));
+          rifWriters.writeValues(CARRIER.class, fieldValues);
+        }
+
+        if (lineNum == 1) {
+          // If lineNum still equals 1, then no line items were successfully added.
+          // Add a single top-level entry.
+          fieldValues.put(CARRIER.LINE_NUM, Integer.toString(lineNum));
+          fieldValues.put(CARRIER.LINE_BENE_PTB_DDCTBL_AMT,
+                  String.format("%.2f", encounter.claim.getTotalDeductiblePaid()));
+          fieldValues.put(CARRIER.LINE_COINSRNC_AMT,
+                  String.format("%.2f", encounter.claim.getTotalCoinsurancePaid()));
+          fieldValues.put(CARRIER.LINE_SBMTD_CHRG_AMT,
+                  String.format("%.2f", encounter.claim.getTotalClaimCost()));
+          fieldValues.put(CARRIER.LINE_ALOWD_CHRG_AMT,
+                  String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+          // Like NCH_CLM_BENE_PMT_AMT, LINE_BENE_PMT_AMT is always zero
+          // (set in field value spreadsheet)
+          fieldValues.put(CARRIER.LINE_PRVDR_PMT_AMT,
+                  String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+          fieldValues.put(CARRIER.LINE_NCH_PMT_AMT,
+                  String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+          // 99241: "Office consultation for a new or established patient"
+          fieldValues.put(CARRIER.HCPCS_CD, "99241");
+          rifWriters.writeValues(CARRIER.class, fieldValues);
+        }
+      }
+      claimCount++;
+    }
+    return claimCount;
+  }
+
+
+  private String getCarrier(String state, CARRIER column) {
+    for (LinkedHashMap<String, String> row : carrierLookup) {
+      if (row.get("STATE").equals(state) || row.get("STATE_CODE").equals(state)) {
+        return row.get(column.toString());
+      }
+    }
+    return "0";
+  }
+
+  /**
+   * Get the Line Place of Service Code. This is a required field for
+   * Carrier and DME claims.
+   * @param encounter The encounter.
+   * @return non-null place of service code.
+   */
+  private String getPlaceOfService(Encounter encounter) {
+    String placeOfServiceCode = "11"; // Default to "11" = Office
+    if (encounter.type.equalsIgnoreCase(EncounterType.VIRTUAL.toString())) {
+      placeOfServiceCode = "02"; // telehealth
+    } else if (encounter.provider.type == ProviderType.IHS) {
+      if (encounter.provider.hasService(EncounterType.WELLNESS)) {
+        placeOfServiceCode = "05"; // no hospitalization
+      } else {
+        placeOfServiceCode = "06"; // hospitalization
+      }
+    } else if (encounter.provider.type == ProviderType.VETERAN) {
+      placeOfServiceCode = "26"; // military
+    } else if (encounter.provider.hasService(EncounterType.SNF)) {
+      placeOfServiceCode = "31"; // skilled nursing facility
+    } else if (encounter.provider.hasService(EncounterType.HOSPICE)) {
+      placeOfServiceCode = "34"; // hospice
+    } else if (encounter.provider.hasService(EncounterType.HOME)) {
+      placeOfServiceCode = "12"; // home
+    } else if (encounter.provider.hasService(EncounterType.URGENTCARE)) {
+      placeOfServiceCode = "20"; // urgent care
+    } else if (encounter.type.equalsIgnoreCase(EncounterType.EMERGENCY.toString())) {
+      placeOfServiceCode = "23"; // emergency room
+    } else if (encounter.type.equalsIgnoreCase(EncounterType.INPATIENT.toString())) {
+      placeOfServiceCode = "21"; // inpatient
+    } else if (encounter.type.equalsIgnoreCase(EncounterType.OUTPATIENT.toString())) {
+      placeOfServiceCode = "22"; // outpatient
+    } else if (encounter.type.equalsIgnoreCase(EncounterType.AMBULATORY.toString())) {
+      placeOfServiceCode = "22";
+    } else if (encounter.type.equalsIgnoreCase(EncounterType.WELLNESS.toString())) {
+      placeOfServiceCode = "11"; // office
+    }
+    return placeOfServiceCode;
+  }
+
+
+  /**
+   * Export prescription claims details for a single person.
+   * @param person the person to export
+   * @param startTime earliest claim date to export
+   * @param stopTime end time of simulation
+   * @return count of claims exported
+   * @throws IOException if something goes wrong
+   */
+  private long exportPrescription(Person person, long startTime, long stopTime)
+        throws IOException {
+    long claimCount = 0;
+    PartDContractHistory partDContracts =
+            (PartDContractHistory) person.attributes.get(ExporterBase.BB2_PARTD_CONTRACTS);
+    // Build a chronologically ordered list of prescription fills (including refills where
+    // specified).
+    List<PrescriptionFill> prescriptionFills = new LinkedList<>();
+    for (Encounter encounter : person.record.encounters) {
+      if (encounter.stop < startTime || encounter.stop < claimCutoff) {
+        continue;
+      }
+      if (ExporterBase.isVAorIHS(encounter)) {
+        continue;
+      }
+
+      for (Medication medication : encounter.medications) {
+        if (!medicationCodeMapper.canMap(medication.codes.get(0).code)) {
+          continue; // skip codes that can't be mapped to NDC
+        }
+        long supplyDaysMax = 90; // TBD - 30, 60, 90 day refil schedules?
+        long supplyInterval = supplyDaysMax * 24 * 60 * 60 * 1000;
+        long finishTime = medication.stop == 0L ? stopTime : Long.min(medication.stop, stopTime);
+        String medicationCode = medicationCodeMapper.map(medication.codes.get(0).code, person);
+        long time = medication.start;
+        int fillNo = 1;
+        while (time < finishTime) {
+          PartDContractID partDContractID = partDContracts.getContractID(time);
+          PrescriptionFill fill = new PrescriptionFill(time, encounter, medication,
+                    medicationCode, fillNo, partDContractID, supplyInterval, finishTime);
+          if (partDContractID != null) {
+            prescriptionFills.add(fill);
+          }
+          if (!fill.refillsRemaining()) {
+            break;
+          }
+          time += Long.min((long)fill.days * 24 * 60 * 60 * 1000, supplyInterval);
+          fillNo++;
+        }
+      }
+    }
+    Collections.sort(prescriptionFills);
+
+    // Export each prescription fill to RIF format
+    HashMap<PDE, String> fieldValues = new HashMap<>();
+    BigDecimal costs = Claim.ZERO_CENTS;
+    int costYear = 0;
+    String catastrophicCode = "";
+    for (PrescriptionFill fill: prescriptionFills) {
+
+      long pdeId = ExporterBase.nextPdeId.getAndDecrement();
+      long claimGroupId = ExporterBase.nextClaimGroupId.getAndDecrement();
+
+      fieldValues.clear();
+      staticFieldConfig.setValues(fieldValues, PDE.class, person);
+
+      // The REQUIRED fields
+      fieldValues.put(PDE.PDE_ID, "" + pdeId);
+      fieldValues.put(PDE.CLM_GRP_ID, "" + claimGroupId);
+      fieldValues.put(PDE.BENE_ID, (String) person.attributes.get(ExporterBase.BB2_BENE_ID));
+      fieldValues.put(PDE.SRVC_DT, ExporterBase.bb2DateFromTimestamp(fill.time));
+      fieldValues.put(PDE.SRVC_PRVDR_ID, fill.encounter.provider.cmsProviderNum);
+      fieldValues.put(PDE.PRSCRBR_ID,
+          "" + (9_999_999_999L - fill.encounter.clinician.identifier));
+      fieldValues.put(PDE.RX_SRVC_RFRNC_NUM, "" + pdeId);
+      fieldValues.put(PDE.PROD_SRVC_ID, fill.medicationCode);
+      // The following field was replaced by the PartD contract ID, leaving this here for now
+      // until this is validated
+      // H=hmo, R=ppo, S=stand-alone, E=employer direct, X=limited income
+      // fieldValues.put(PrescriptionFields.PLAN_CNTRCT_REC_ID,
+      //     ("R" + Math.abs(
+      //         UUID.fromString(medication.claim.payer.uuid)
+      //         .getMostSignificantBits())).substring(0, 5));
+      fieldValues.put(PDE.PLAN_CNTRCT_REC_ID, fill.partDContractID.toString());
+      fieldValues.put(PDE.DAW_PROD_SLCTN_CD, "" + (int) person.rand(0, 9));
+      fieldValues.put(PDE.QTY_DSPNSD_NUM, "" + fill.quantity);
+      fieldValues.put(PDE.DAYS_SUPLY_NUM, "" + fill.days);
+      fieldValues.put(PDE.FILL_NUM, "" + fill.fillNo);
+      int year = Utilities.getYear(fill.time);
+      if (year != costYear) {
+        costYear = year;
+        costs = Claim.ZERO_CENTS;
+        catastrophicCode = ""; // Blank = Attachment point not met
+      }
+      BigDecimal threshold = getDrugOutOfPocketThreshold(year);
+      costs = costs.add(fill.medication.claim.getTotalPatientCost());
+      if (costs.compareTo(threshold) < 0) {
+        fieldValues.put(PDE.GDC_BLW_OOPT_AMT, String.format("%.2f", costs));
+        fieldValues.put(PDE.GDC_ABV_OOPT_AMT, "0");
+        fieldValues.put(PDE.CTSTRPHC_CVRG_CD, catastrophicCode);
+      } else {
+        if (catastrophicCode.equals("")) {
+          catastrophicCode = "A"; // A = Attachment point met on this event
+        } else if (catastrophicCode.equals("A")) {
+          catastrophicCode  = "C"; // C = Above attachment point
+        }
+        fieldValues.put(PDE.GDC_BLW_OOPT_AMT, String.format("%.2f", threshold));
+        fieldValues.put(PDE.GDC_ABV_OOPT_AMT,
+                String.format("%.2f", costs.subtract(threshold)));
+        fieldValues.put(PDE.CTSTRPHC_CVRG_CD, catastrophicCode);
+      }
+      fieldValues.put(PDE.TOT_RX_CST_AMT,
+          String.format("%.2f", fill.medication.claim.getTotalClaimCost()));
+      // Under normal circumstances, the following fields summed together,
+      // should equal TOT_RX_CST_AMT:
+      // - PTNT_PAY_AMT       : what the patient paid
+      // - OTHR_TROOP_AMT     : what 3rd party paid out of pocket
+      // - LICS_AMT           : low income subsidized payment
+      // - PLRO_AMT           : what other 3rd party insurances paid
+      // - CVRD_D_PLAN_PD_AMT : what Part D paid
+      // - NCVRD_PLAN_PD_AMT  : part of total not covered by Part D whatsoever
+      // OTHR_TROOP_AMT and LICS_AMT are always 0, set in field value spreadsheet
+      fieldValues.put(PDE.PTNT_PAY_AMT,
+          String.format("%.2f", fill.medication.claim.getTotalPatientCost()));
+      fieldValues.put(PDE.PLRO_AMT,
+          String.format("%.2f", fill.medication.claim.getTotalPaidBySecondaryPayer()));
+      fieldValues.put(PDE.CVRD_D_PLAN_PD_AMT,
+          String.format("%.2f", fill.medication.claim.getTotalCoveredCost()));
+      fieldValues.put(PDE.NCVRD_PLAN_PD_AMT,
+          String.format("%.2f", fill.medication.claim.getTotalAdjustment()));
+
+      fieldValues.put(PDE.PHRMCY_SRVC_TYPE_CD, "0" + (int) person.rand(1, 8));
+      fieldValues.put(PDE.PD_DT, ExporterBase.bb2DateFromTimestamp(fill.time));
+      // 00=not specified, 01=home, 02=SNF, 03=long-term, 11=hospice, 14=homeless
+      if (person.attributes.containsKey("homeless")
+          && ((Boolean) person.attributes.get("homeless") == true)) {
+        fieldValues.put(PDE.PTNT_RSDNC_CD, "14");
+      } else {
+        fieldValues.put(PDE.PTNT_RSDNC_CD, "01");
+      }
+
+      rifWriters.writeValues(PDE.class, fieldValues);
+      claimCount++;
+    }
+    return claimCount;
+  }
+
+  protected BigDecimal getDrugOutOfPocketThreshold(int year) {
+    double threshold = pdeOutOfPocketThresholds.getOrDefault(year, 4550.0);
+    return BigDecimal.valueOf(threshold);
+  }
+
+
+  /**
+   * Export DME details for a single person.
+   * @param person the person to export
+   * @param startTime earliest claim date to export
+   * @param stopTime end time of simulation
+   * @return count of claims exported
+   * @throws IOException if something goes wrong
+   */
+  private long exportDME(Person person, long startTime, long stopTime)
+        throws IOException {
+    long claimCount = 0;
+    HashMap<DME, String> fieldValues = new HashMap<>();
+
+    for (Encounter encounter : person.record.encounters) {
+      if (encounter.stop < startTime || encounter.stop < claimCutoff) {
+        continue;
+      }
+      if (ExporterBase.isVAorIHS(encounter)) {
+        continue;
+      }
+
+      long claimId = ExporterBase.nextClaimId.getAndDecrement();
+      long claimGroupId = ExporterBase.nextClaimGroupId.getAndDecrement();
+      long carrClmId = ExporterBase.nextCarrClmCntlNum.getAndDecrement();
+
+      double latestHemoglobin = 0;
+      for (HealthRecord.Observation observation : encounter.observations) {
+        if (observation.containsCode("718-7", "http://loinc.org")) {
+          latestHemoglobin = (double) observation.value;
+        }
+      }
+
+      fieldValues.clear();
+      staticFieldConfig.setValues(fieldValues, DME.class, person);
+
+      // complex fields that could not easily be set using cms_field_values.tsv
+      fieldValues.put(DME.CLM_ID, "" + claimId);
+      fieldValues.put(DME.CLM_GRP_ID, "" + claimGroupId);
+      fieldValues.put(DME.CARR_CLM_CNTL_NUM, "" + carrClmId);
+      fieldValues.put(DME.BENE_ID, (String) person.attributes.get(ExporterBase.BB2_BENE_ID));
+      fieldValues.put(DME.LINE_HCT_HGB_RSLT_NUM, "" + latestHemoglobin);
+      fieldValues.put(DME.CARR_NUM,
+              getCarrier(encounter.provider.state, CARRIER.CARR_NUM));
+      fieldValues.put(DME.NCH_WKLY_PROC_DT,
+              ExporterBase.bb2DateFromTimestamp(ExportHelper.nextFriday(encounter.stop)));
+      fieldValues.put(DME.PRVDR_NUM, encounter.provider.cmsProviderNum);
+      fieldValues.put(DME.PRVDR_NPI, encounter.provider.npi);
+      fieldValues.put(DME.RFR_PHYSN_NPI, encounter.clinician.npi);
+      fieldValues.put(DME.RFR_PHYSN_UPIN, encounter.provider.cmsUpin);
+      fieldValues.put(DME.PRVDR_SPCLTY,
+          ClinicianSpecialty.getCMSProviderSpecialtyCode(
+              (String) encounter.clinician.attributes.get(Clinician.SPECIALTY)));
+      fieldValues.put(DME.PRVDR_STATE_CD,
+              locationMapper.getStateCode(encounter.provider.state));
+      fieldValues.put(DME.TAX_NUM,
+              BeneficiaryExporter.bb2TaxId((String)encounter.clinician.attributes.get(Person.IDENTIFIER_SSN)));
+      fieldValues.put(DME.DMERC_LINE_PRCNG_STATE_CD,
+              locationMapper.getStateCode((String)person.attributes.get(Person.STATE)));
+      fieldValues.put(DME.LINE_1ST_EXPNS_DT, ExporterBase.bb2DateFromTimestamp(encounter.start));
+      fieldValues.put(DME.LINE_LAST_EXPNS_DT, ExporterBase.bb2DateFromTimestamp(encounter.stop));
+      fieldValues.put(DME.LINE_SRVC_CNT, "" + encounter.claim.items.size());
+      fieldValues.put(DME.LINE_PLACE_OF_SRVC_CD, getPlaceOfService(encounter));
+
+      // OPTIONAL
+      if (encounter.reason != null) {
+        // If the encounter has a recorded reason, enter the mapped
+        // values into the principle diagnoses code.
+        if (conditionCodeMapper.canMap(encounter.reason.code)) {
+          String icdCode = conditionCodeMapper.map(encounter.reason.code, person, true);
+          fieldValues.put(DME.PRNCPAL_DGNS_CD, icdCode);
+          fieldValues.put(DME.LINE_ICD_DGNS_CD, icdCode);
+        }
+      }
+
+      // Use the active condition diagnoses to enter mapped values
+      // into the diagnoses codes.
+      List<String> mappedDiagnosisCodes = getDiagnosesCodes(person, encounter.stop);
+      if (mappedDiagnosisCodes.isEmpty()) {
+        continue; // skip this encounter
+      }
+      int smallest = Math.min(mappedDiagnosisCodes.size(), BB2RIFStructure.dmeDxFields.length);
+      for (int i = 0; i < smallest; i++) {
+        DME[] dxField = BB2RIFStructure.dmeDxFields[i];
+        fieldValues.put(dxField[0], mappedDiagnosisCodes.get(i));
+        fieldValues.put(dxField[1], "0"); // 0=ICD10
+      }
+      if (!fieldValues.containsKey(DME.PRNCPAL_DGNS_CD)) {
+        fieldValues.put(DME.PRNCPAL_DGNS_CD, mappedDiagnosisCodes.get(0));
+        fieldValues.put(DME.LINE_ICD_DGNS_CD, mappedDiagnosisCodes.get(0));
+      }
+
+      // preprocess some subtotals...
+      ClaimEntry subTotals = (encounter.claim).new ClaimEntry(null);
+      for (ClaimEntry lineItem : encounter.claim.items) {
+        if (lineItem.entry instanceof Device || lineItem.entry instanceof Supply) {
+          subTotals.addCosts(lineItem);
+        }
+      }
+      fieldValues.put(DME.CARR_CLM_CASH_DDCTBL_APLD_AMT,
+          String.format("%.2f", subTotals.deductiblePaidByPatient));
+      fieldValues.put(DME.NCH_CARR_CLM_SBMTD_CHRG_AMT,
+          String.format("%.2f", subTotals.cost));
+      BigDecimal paidAmount = subTotals.coinsurancePaidByPayer.add(subTotals.paidByPayer);
+      fieldValues.put(DME.CARR_CLM_PRMRY_PYR_PD_AMT,
+          String.format("%.2f", paidAmount));
+      fieldValues.put(DME.NCH_CARR_CLM_ALOWD_AMT,
+          String.format("%.2f", paidAmount));
+      fieldValues.put(DME.NCH_CLM_PRVDR_PMT_AMT,
+          String.format("%.2f", paidAmount));
+      fieldValues.put(DME.CLM_PMT_AMT,
+          String.format("%.2f", paidAmount));
+
+      synchronized (rifWriters.getOrCreateWriter(DME.class)) {
+        int lineNum = 1;
+        boolean wroteAtLeastOneLine = false;
+        // Now generate the line items...
+        for (ClaimEntry lineItem : encounter.claim.items) {
+          if (!(lineItem.entry instanceof Device || lineItem.entry instanceof Supply)) {
+            continue;
+          }
+          if (lineItem.entry instanceof Supply) {
+            Supply supply = (Supply) lineItem.entry;
+            fieldValues.put(DME.DMERC_LINE_MTUS_CNT, "" + supply.quantity);
+          } else {
+            fieldValues.put(DME.DMERC_LINE_MTUS_CNT, "");
+          }
+          if (!dmeCodeMapper.canMap(lineItem.entry.codes.get(0).code)) {
+            System.err.println(" *** Possibly Missing DME Code: "
+                + lineItem.entry.codes.get(0).code
+                + " " + lineItem.entry.codes.get(0).display);
+            continue;
+          }
+          fieldValues.put(DME.CLM_FROM_DT, ExporterBase.bb2DateFromTimestamp(lineItem.entry.start));
+          fieldValues.put(DME.CLM_THRU_DT, ExporterBase.bb2DateFromTimestamp(lineItem.entry.start));
+          String hcpcsCode = dmeCodeMapper.map(lineItem.entry.codes.get(0).code, person);
+          fieldValues.put(DME.HCPCS_CD, hcpcsCode);
+          if (betosCodeMapper.canMap(hcpcsCode)) {
+            fieldValues.put(DME.BETOS_CD, betosCodeMapper.map(hcpcsCode, person));
+          } else {
+            fieldValues.put(DME.BETOS_CD, "");
+          }
+          fieldValues.put(DME.LINE_CMS_TYPE_SRVC_CD,
+                  dmeCodeMapper.map(lineItem.entry.codes.get(0).code,
+                          DME.LINE_CMS_TYPE_SRVC_CD.toString().toLowerCase(),
+                          person));
+          fieldValues.put(DME.LINE_BENE_PTB_DDCTBL_AMT,
+                  String.format("%.2f", lineItem.deductiblePaidByPatient));
+          fieldValues.put(DME.LINE_COINSRNC_AMT,
+                  String.format("%.2f", lineItem.getCoinsurancePaid()));
+          // LINE_BENE_PMT_AMT and NCH_CLM_BENE_PMT_AMT are always 0, set in field value spreadsheet
+          BigDecimal providerAmount = lineItem.coinsurancePaidByPayer.add(lineItem.paidByPayer);
+          fieldValues.put(DME.LINE_PRVDR_PMT_AMT,
+              String.format("%.2f", providerAmount));
+          fieldValues.put(DME.LINE_NCH_PMT_AMT,
+              String.format("%.2f", providerAmount));
+          fieldValues.put(DME.LINE_SBMTD_CHRG_AMT,
+              String.format("%.2f", lineItem.cost));
+          BigDecimal allowedAmount = lineItem.cost.subtract(lineItem.adjustment);
+          fieldValues.put(DME.LINE_ALOWD_CHRG_AMT,
+              String.format("%.2f", allowedAmount));
+          fieldValues.put(DME.LINE_PRMRY_ALOWD_CHRG_AMT,
+              String.format("%.2f", allowedAmount));
+
+          // set the line number and write out field values
+          fieldValues.put(DME.LINE_NUM, Integer.toString(lineNum++));
+          rifWriters.writeValues(DME.class, fieldValues);
+          wroteAtLeastOneLine = true;
+        }
+        if (wroteAtLeastOneLine) {
+          claimCount++;
+        }
+      }
+    }
+    return claimCount;
+  }
+
+  /**
+   * Export Home Health Agency visits for a single person.
+   * @param person the person to export
+   * @param startTime earliest claim date to export
+   * @param stopTime end time of simulation
+   * @return count of claims exported
+   * @throws IOException if something goes wrong
+   */
+  private long exportHome(Person person, long startTime, long stopTime) throws IOException {
+    HashMap<HHA, String> fieldValues = new HashMap<>();
+    long claimCount = 0;
+    int homeVisits = 0;
+
+    long maxGapForContinuousHHAService = Utilities.convertTime("days", 2);
+    ConsolidatedServicePeriods servicePeriods = new ConsolidatedServicePeriods(
+            maxGapForContinuousHHAService);
+    for (Encounter encounter : person.record.encounters) {
+      if (encounter.stop < startTime || encounter.stop < claimCutoff) {
+        continue;
+      }
+      if (!ExporterBase.getClaimTypes(encounter).contains(ClaimType.HHA)) {
+        continue;
+      }
+      List<String> mappedDiagnosisCodes = getDiagnosesCodes(person, encounter.stop);
+      if (mappedDiagnosisCodes.isEmpty()) {
+        continue; // skip this encounter
+      }
+      servicePeriods.addEncounter(encounter);
+    }
+
+    for (ConsolidatedServicePeriod servicePeriod: servicePeriods.getPeriods()) {
+      long claimId = ExporterBase.nextClaimId.getAndDecrement();
+      long claimGroupId = ExporterBase.nextClaimGroupId.getAndDecrement();
+      long fiDocId = ExporterBase.nextFiDocCntlNum.getAndDecrement();
+
+      fieldValues.clear();
+      staticFieldConfig.setValues(fieldValues, HHA.class, person);
+
+      // The REQUIRED fields
+      fieldValues.put(HHA.BENE_ID,  (String) person.attributes.get(ExporterBase.BB2_BENE_ID));
+      fieldValues.put(HHA.CLM_ID, "" + claimId);
+      fieldValues.put(HHA.CLM_GRP_ID, "" + claimGroupId);
+      fieldValues.put(HHA.FI_DOC_CLM_CNTL_NUM, "" + fiDocId);
+      fieldValues.put(HHA.CLM_FROM_DT, ExporterBase.bb2DateFromTimestamp(servicePeriod.getStart()));
+      fieldValues.put(HHA.CLM_ADMSN_DT, ExporterBase.bb2DateFromTimestamp(servicePeriod.getStart()));
+      fieldValues.put(HHA.CLM_THRU_DT, ExporterBase.bb2DateFromTimestamp(servicePeriod.getStop()));
+      fieldValues.put(HHA.NCH_WKLY_PROC_DT,
+          ExporterBase.bb2DateFromTimestamp(ExportHelper.nextFriday(servicePeriod.getStop())));
+
+      // random from fields TSV, may be overriden below
+      String revCenter = fieldValues.get(HHA.REV_CNTR);
+
+      final String HHA_TOTAL_CHARGE_REV_CNTR = "0001"; // Total charge
+      final String HHA_GENERAL_REV_CNTR = "0270"; // General medical/surgical supplies
+      final String HHA_MEDICATION_CODE = "T1502"; // Administration of medication
+      ConsolidatedClaimLines consolidatedClaimLines = new ConsolidatedClaimLines();
+      for (Encounter encounter : servicePeriod.getEncounters()) {
+        for (ClaimEntry lineItem : encounter.claim.items) {
+          String hcpcsCode = null;
+          if (lineItem.entry instanceof HealthRecord.Procedure) {
+            for (HealthRecord.Code code : lineItem.entry.codes) {
+              if (hcpcsCodeMapper.canMap(code.code)) {
+                hcpcsCode = hcpcsCodeMapper.map(code.code, person, true);
+                if (hhaRevCntrMapper.canMap(code.code)) {
+                  revCenter = hhaRevCntrMapper.map(code.code, person);
+                }
+                break; // take the first mappable code for each procedure
+              }
+            }
+            if (hcpcsCode == null) {
+              revCenter = HHA_GENERAL_REV_CNTR;
+            }
+            consolidatedClaimLines.addClaimLine(hcpcsCode, revCenter, lineItem, encounter);
+          } else if (lineItem.entry instanceof HealthRecord.Medication) {
+            HealthRecord.Medication med = (HealthRecord.Medication) lineItem.entry;
+            if (med.administration) {
+              hcpcsCode = HHA_MEDICATION_CODE;
+              revCenter = HHA_GENERAL_REV_CNTR;
+              consolidatedClaimLines.addClaimLine(hcpcsCode, revCenter, lineItem, encounter);
+            }
+          }
+        }
+      }
+
+      fieldValues.put(HHA.CLM_PMT_AMT,
+          String.format("%.2f", consolidatedClaimLines.getCoveredCost()));
+      fieldValues.put(HHA.CLM_TOT_CHRG_AMT,
+          String.format("%.2f", consolidatedClaimLines.getTotalClaimCost()));
+      fieldValues.put(HHA.CLM_HHA_TOT_VISIT_CNT, "" + servicePeriod.getEncounters().size());
+      fieldValues.put(HHA.PRVDR_NUM, servicePeriod.getProvider().cmsProviderNum);
+      fieldValues.put(HHA.ORG_NPI_NUM, servicePeriod.getProvider().npi);
+      fieldValues.put(HHA.PRVDR_STATE_CD,
+          locationMapper.getStateCode(servicePeriod.getProvider().state));
+
+      // Use the final encounter in the service period to set all of the remaining field values that
+      // are the same for all claim lines
+      // TODO: update ConsolidatedServicePeriods to separate encounters based on provider, clinician
+      Encounter finalEncounterOfPeriod = servicePeriod.getEncounters().get(
+              servicePeriod.getEncounters().size() - 1);
+      if (finalEncounterOfPeriod.claim.plan == PayerManager
+              .getGovernmentPayer(PayerManager.MEDICARE)
+              .getGovernmentPayerPlan()) {
+        fieldValues.put(HHA.NCH_PRMRY_PYR_CLM_PD_AMT, "0");
+      } else {
+        fieldValues.put(HHA.NCH_PRMRY_PYR_CLM_PD_AMT,
+            String.format("%.2f", servicePeriod.getTotalCost().getCoveredCost()));
+      }
+      if (finalEncounterOfPeriod.reason != null) {
+        // If the encounter has a recorded reason, enter the mapped
+        // values into the principle diagnoses code.
+        if (conditionCodeMapper.canMap(finalEncounterOfPeriod.reason.code)) {
+          String icdCode = conditionCodeMapper.map(finalEncounterOfPeriod.reason.code, person,
+                  true);
+          fieldValues.put(HHA.PRNCPAL_DGNS_CD, icdCode);
+        }
+      }
+      // PTNT_DSCHRG_STUS_CD: 1=home, 2=transfer, 3=SNF, 20=died, 30=still here
+      String dischargeStatus = "1"; //discharged
+      if (!person.alive(servicePeriod.getStop())) {
+        dischargeStatus = "20"; // the patient died before the service period ended
+      } else if (servicePeriod.getStop() > stopTime) {
+        // TBD: revisit if we break up service periods into multiple billing periods, can then set
+        // this for all but the final billing period within a service period.
+        dischargeStatus = "30"; // the patient is still having treatment
+      }
+      fieldValues.put(HHA.PTNT_DSCHRG_STUS_CD, dischargeStatus);
+      // Use the active condition diagnoses to enter mapped values
+      // into the diagnoses codes.
+      List<String> mappedDiagnosisCodes = getDiagnosesCodes(person, finalEncounterOfPeriod.stop);
+      int smallest = Math.min(mappedDiagnosisCodes.size(),
+              BB2RIFStructure.homeDxFields.length);
+      for (int i = 0; i < smallest; i++) {
+        HHA[] dxField = BB2RIFStructure.homeDxFields[i];
+        fieldValues.put(dxField[0], mappedDiagnosisCodes.get(i));
+        fieldValues.put(dxField[1], "0"); // 0=ICD10
+      }
+      if (!fieldValues.containsKey(HHA.PRNCPAL_DGNS_CD)) {
+        fieldValues.put(HHA.PRNCPAL_DGNS_CD, mappedDiagnosisCodes.get(0));
+      }
+
+      // Check for external code...
+      setExternalCode(person, fieldValues,
+          HHA.PRNCPAL_DGNS_CD, HHA.ICD_DGNS_E_CD1, HHA.ICD_DGNS_E_VRSN_CD1);
+      setExternalCode(person, fieldValues,
+          HHA.PRNCPAL_DGNS_CD, HHA.FST_DGNS_E_CD, HHA.FST_DGNS_E_VRSN_CD);
+
+      // now loop over all of the consolidated claim lines and write a row for each
+      synchronized (rifWriters.getOrCreateWriter(HHA.class)) {
+        int claimLine = 1;
+        for (ConsolidatedClaimLines.ConsolidatedClaimLine lineItem:
+                consolidatedClaimLines.getLines()) {
+          fieldValues.put(HHA.HCPCS_CD, lineItem.getCode());
+          fieldValues.put(HHA.AT_PHYSN_NPI, lineItem.getClinician().npi);
+          fieldValues.put(HHA.RNDRNG_PHYSN_NPI, lineItem.getClinician().npi);
+          fieldValues.put(HHA.REV_CNTR_DT, ExporterBase.bb2DateFromTimestamp(lineItem.getStart()));
+          int revCntrCount = lineItem.getCount();
+          switch (lineItem.getCode()) {
+            case HHA_MEDICATION_CODE:
+              fieldValues.put(HHA.REV_CNTR, lineItem.getRevCntr());
+              fieldValues.put(HHA.REV_CNTR_NDC_QTY, Integer.toString(revCntrCount));
+              fieldValues.put(HHA.REV_CNTR_NDC_QTY_QLFR_CD, "UN"); // Unit
+              fieldValues.remove(HHA.REV_CNTR_UNIT_CNT);
+              break;
+            default:
+              fieldValues.put(HHA.REV_CNTR, lineItem.getRevCntr());
+              fieldValues.put(HHA.REV_CNTR_UNIT_CNT, Integer.toString(revCntrCount));
+              fieldValues.remove(HHA.REV_CNTR_NDC_QTY);
+              fieldValues.remove(HHA.REV_CNTR_NDC_QTY_QLFR_CD);
+              break;
+          }
+
+          fieldValues.put(HHA.CLM_LINE_NUM, Integer.toString(claimLine++));
+          setHHAClaimLineCosts(fieldValues, lineItem, Integer.max(1, revCntrCount));
+          rifWriters.writeValues(HHA.class, fieldValues);
+        }
+
+        // Add a total charge entry.
+        setHHAClaimLineCosts(fieldValues, consolidatedClaimLines, 1);
+        fieldValues.put(HHA.CLM_LINE_NUM, Integer.toString(claimLine++));
+        fieldValues.put(HHA.REV_CNTR, HHA_TOTAL_CHARGE_REV_CNTR);
+        fieldValues.put(HHA.REV_CNTR_DT, ExporterBase.bb2DateFromTimestamp(servicePeriod.getStop()));
+        fieldValues.remove(HHA.HCPCS_CD);
+        fieldValues.remove(HHA.AT_PHYSN_NPI);
+        fieldValues.remove(HHA.RNDRNG_PHYSN_NPI);
+        fieldValues.remove(HHA.REV_CNTR_RATE_AMT);
+        fieldValues.remove(HHA.REV_CNTR_UNIT_CNT);
+        fieldValues.remove(HHA.REV_CNTR_DDCTBL_COINSRNC_CD);
+        rifWriters.writeValues(HHA.class, fieldValues);
+      }
+      claimCount++;
+    }
+    return claimCount;
+  }
+
+  private static void setHHAClaimLineCosts(HashMap<HHA, String> fieldValues, ClaimCost lineItem,
+          int count) {
+    fieldValues.put(HHA.REV_CNTR_RATE_AMT,
+            String.format("%.2f", lineItem.cost
+                    .divide(BigDecimal.valueOf(count), RoundingMode.HALF_EVEN)
+                    .setScale(2, RoundingMode.HALF_EVEN)));
+    fieldValues.put(HHA.REV_CNTR_PMT_AMT_AMT,
+            String.format("%.2f", lineItem.coinsurancePaidByPayer.add(lineItem.paidByPayer)));
+    fieldValues.put(HHA.REV_CNTR_TOT_CHRG_AMT,
+            String.format("%.2f", lineItem.cost));
+    fieldValues.put(HHA.REV_CNTR_NCVRD_CHRG_AMT,
+            String.format("%.2f", lineItem.copayPaidByPatient
+                    .add(lineItem.deductiblePaidByPatient).add(lineItem.patientOutOfPocket)));
+    if (lineItem.patientOutOfPocket.compareTo(Claim.ZERO_CENTS) == 0
+            && lineItem.deductiblePaidByPatient.compareTo(Claim.ZERO_CENTS) == 0) {
+      // Not subject to deductible or coinsurance
+      fieldValues.put(HHA.REV_CNTR_DDCTBL_COINSRNC_CD, "3");
+    } else if (lineItem.patientOutOfPocket.compareTo(Claim.ZERO_CENTS) > 0
+            && lineItem.deductiblePaidByPatient.compareTo(Claim.ZERO_CENTS) > 0) {
+      // Subject to deductible and coinsurance
+      fieldValues.put(HHA.REV_CNTR_DDCTBL_COINSRNC_CD, "0");
+    } else if (lineItem.patientOutOfPocket.compareTo(Claim.ZERO_CENTS) == 0) {
+      // Not subject to deductible
+      fieldValues.put(HHA.REV_CNTR_DDCTBL_COINSRNC_CD, "1");
+    } else {
+      // Not subject to coinsurance
+      fieldValues.put(HHA.REV_CNTR_DDCTBL_COINSRNC_CD, "2");
+    }
+  }
+
+  /**
+   * Export Home Health Agency visits for a single person.
+   * @param person the person to export
+   * @param startTime earliest claim date to export
+   * @param stopTime end time of simulation
+   * @return count of claims exported
+   * @throws IOException if something goes wrong
+   */
+  private long exportHospice(Person person, long startTime, long stopTime) throws IOException {
+    long claimCount = 0;
+    HashMap<HOSPICE, String> fieldValues = new HashMap<>();
+    for (Encounter encounter : person.record.encounters) {
+      if (encounter.stop < startTime || encounter.stop < claimCutoff) {
+        continue;
+      }
+      if (!ExporterBase.getClaimTypes(encounter).contains(ClaimType.HOSPICE)) {
+        continue;
+      }
+
+      // Use the active condition diagnoses to enter mapped values
+      // into the diagnoses codes.
+      List<String> mappedDiagnosisCodes = getDiagnosesCodes(person, encounter.stop);
+      if (mappedDiagnosisCodes.isEmpty()) {
+        continue; // skip this encounter
+      }
+
+      long claimId = ExporterBase.nextClaimId.getAndDecrement();
+      long claimGroupId = ExporterBase.nextClaimGroupId.getAndDecrement();
+      long fiDocId = ExporterBase.nextFiDocCntlNum.getAndDecrement();
+      fieldValues.clear();
+      staticFieldConfig.setValues(fieldValues, HOSPICE.class, person);
+
+      fieldValues.put(HOSPICE.BENE_ID, (String) person.attributes.get(ExporterBase.BB2_BENE_ID));
+      fieldValues.put(HOSPICE.CLM_ID, "" + claimId);
+      fieldValues.put(HOSPICE.CLM_GRP_ID, "" + claimGroupId);
+      fieldValues.put(HOSPICE.FI_DOC_CLM_CNTL_NUM, "" + fiDocId);
+      fieldValues.put(HOSPICE.CLM_FROM_DT, ExporterBase.bb2DateFromTimestamp(encounter.start));
+      fieldValues.put(HOSPICE.CLM_HOSPC_START_DT_ID, ExporterBase.bb2DateFromTimestamp(encounter.start));
+      fieldValues.put(HOSPICE.CLM_THRU_DT, ExporterBase.bb2DateFromTimestamp(encounter.stop));
+      fieldValues.put(HOSPICE.NCH_WKLY_PROC_DT,
+              ExporterBase.bb2DateFromTimestamp(ExportHelper.nextFriday(encounter.stop)));
+      fieldValues.put(HOSPICE.PRVDR_NUM, encounter.provider.cmsProviderNum);
+      fieldValues.put(HOSPICE.AT_PHYSN_NPI, encounter.clinician.npi);
+      fieldValues.put(HOSPICE.RNDRNG_PHYSN_NPI, encounter.clinician.npi);
+      fieldValues.put(HOSPICE.ORG_NPI_NUM, encounter.provider.npi);
+      fieldValues.put(HOSPICE.CLM_PMT_AMT,
+              String.format("%.2f", encounter.claim.getTotalClaimCost()));
+      if (encounter.claim.plan == PayerManager.getGovernmentPayer(PayerManager.MEDICARE)
+          .getGovernmentPayerPlan()) {
+        fieldValues.put(HOSPICE.NCH_PRMRY_PYR_CLM_PD_AMT, "0");
+      } else {
+        fieldValues.put(HOSPICE.NCH_PRMRY_PYR_CLM_PD_AMT,
+                String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+      }
+      fieldValues.put(HOSPICE.PRVDR_STATE_CD,
+              locationMapper.getStateCode(encounter.provider.state));
+      // PTNT_DSCHRG_STUS_CD: 1=home, 2=transfer, 3=SNF, 20=died, 30=still here
+      // NCH_PTNT_STUS_IND_CD: A = Discharged, B = Died, C = Still a patient
+      String dischargeStatus = null;
+      String patientStatus = null;
+      String dischargeDate = null;
+      if (encounter.ended) {
+        dischargeStatus = "1"; // TODO 2=transfer if the next encounter is also inpatient
+        patientStatus = "A"; // discharged
+        dischargeDate = ExporterBase.bb2DateFromTimestamp(ExportHelper.nextFriday(encounter.stop));
+      } else {
+        dischargeStatus = "30"; // the patient is still here
+        patientStatus = "C"; // still a patient
+      }
+      if (!person.alive(encounter.stop)) {
+        dischargeStatus = "20"; // the patient died before the encounter ended
+        patientStatus = "B"; // died
+        dischargeDate = ExporterBase.bb2DateFromTimestamp(ExportHelper.nextFriday(encounter.stop));
+      }
+      fieldValues.put(HOSPICE.PTNT_DSCHRG_STUS_CD, dischargeStatus);
+      fieldValues.put(HOSPICE.NCH_PTNT_STATUS_IND_CD, patientStatus);
+      if (dischargeDate != null) {
+        fieldValues.put(HOSPICE.NCH_BENE_DSCHRG_DT, dischargeDate);
+      }
+      fieldValues.put(HOSPICE.CLM_TOT_CHRG_AMT,
+              String.format("%.2f", encounter.claim.getTotalClaimCost()));
+      fieldValues.put(HOSPICE.REV_CNTR_TOT_CHRG_AMT,
+          String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+      fieldValues.put(HOSPICE.REV_CNTR_NCVRD_CHRG_AMT,
+          String.format("%.2f", encounter.claim.getTotalPatientCost()));
+      fieldValues.put(HOSPICE.REV_CNTR_PMT_AMT_AMT,
+          String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+      fieldValues.put(HOSPICE.REV_CNTR_PRVDR_PMT_AMT,
+              String.format("%.2f", encounter.claim.getTotalCoveredCost()));
+
+      if (encounter.reason != null) {
+        // If the encounter has a recorded reason, enter the mapped
+        // values into the principle diagnoses code.
+        if (conditionCodeMapper.canMap(encounter.reason.code)) {
+          String icdCode = conditionCodeMapper.map(encounter.reason.code, person, true);
+          fieldValues.put(HOSPICE.PRNCPAL_DGNS_CD, icdCode);
+        }
+      }
+
+      int smallest = Math.min(mappedDiagnosisCodes.size(),
+              BB2RIFStructure.hospiceDxFields.length);
+      for (int i = 0; i < smallest; i++) {
+        HOSPICE[] dxField = BB2RIFStructure.hospiceDxFields[i];
+        fieldValues.put(dxField[0], mappedDiagnosisCodes.get(i));
+        fieldValues.put(dxField[1], "0"); // 0=ICD10
+      }
+      if (!fieldValues.containsKey(HOSPICE.PRNCPAL_DGNS_CD)) {
+        fieldValues.put(HOSPICE.PRNCPAL_DGNS_CD, mappedDiagnosisCodes.get(0));
+      }
+
+      // Check for external code...
+      setExternalCode(person, fieldValues,
+          HOSPICE.PRNCPAL_DGNS_CD, HOSPICE.ICD_DGNS_E_CD1, HOSPICE.ICD_DGNS_E_VRSN_CD1);
+      setExternalCode(person, fieldValues,
+          HOSPICE.PRNCPAL_DGNS_CD, HOSPICE.FST_DGNS_E_CD, HOSPICE.FST_DGNS_E_VRSN_CD);
+
+      int days = (int) ((encounter.stop - encounter.start) / (1000 * 60 * 60 * 24));
+      if (days <= 0) {
+        days = 1;
+      }
+      fieldValues.put(HOSPICE.CLM_UTLZTN_DAY_CNT, "" + days);
+      int coinDays = days -  21; // first 21 days no coinsurance
+      if (coinDays < 0) {
+        coinDays = 0;
+      }
+      fieldValues.put(HOSPICE.REV_CNTR_UNIT_CNT, "" + days);
+      fieldValues.put(HOSPICE.REV_CNTR_RATE_AMT,
+          String.format("%.2f", encounter.claim.getTotalClaimCost()
+                  .divide(BigDecimal.valueOf(days), RoundingMode.HALF_EVEN)
+                  .setScale(2, RoundingMode.HALF_EVEN)));
+      String revCenter = fieldValues.get(HOSPICE.REV_CNTR);
+
+      synchronized (rifWriters.getOrCreateWriter(HOSPICE.class)) {
+        int claimLine = 1;
+        for (ClaimEntry lineItem : encounter.claim.items) {
+          String hcpcsCode = null;
+          if (lineItem.entry instanceof HealthRecord.Procedure) {
+            for (HealthRecord.Code code : lineItem.entry.codes) {
+              if (hcpcsCodeMapper.canMap(code.code)) {
+                hcpcsCode = hcpcsCodeMapper.map(code.code, person, true);
+                break; // take the first mappable code for each procedure
+              }
+            }
+            fieldValues.put(HOSPICE.REV_CNTR, revCenter);
+            fieldValues.remove(HOSPICE.REV_CNTR_NDC_QTY);
+            fieldValues.remove(HOSPICE.REV_CNTR_NDC_QTY_QLFR_CD);
+          } else if (lineItem.entry instanceof HealthRecord.Medication) {
+            HealthRecord.Medication med = (HealthRecord.Medication) lineItem.entry;
+            if (med.administration) {
+              hcpcsCode = "T1502";  // Administration of medication
+              fieldValues.put(HOSPICE.REV_CNTR, "0250"); // Pharmacy-general classification
+              fieldValues.put(HOSPICE.REV_CNTR_NDC_QTY, "1"); // 1 Unit
+              fieldValues.put(HOSPICE.REV_CNTR_NDC_QTY_QLFR_CD, "UN"); // Unit
+            }
+          }
+          if (hcpcsCode == null) {
+            continue;
+          }
+
+          fieldValues.put(HOSPICE.CLM_LINE_NUM, Integer.toString(claimLine++));
+          fieldValues.put(HOSPICE.REV_CNTR_DT, ExporterBase.bb2DateFromTimestamp(lineItem.entry.start));
+          fieldValues.put(HOSPICE.HCPCS_CD, hcpcsCode);
+          fieldValues.put(HOSPICE.REV_CNTR_RATE_AMT,
+              String.format("%.2f", lineItem.cost
+                      .divide(BigDecimal.valueOf(Integer.max(1, days)), RoundingMode.HALF_EVEN)
+                      .setScale(2, RoundingMode.HALF_EVEN)));
+          fieldValues.put(HOSPICE.REV_CNTR_PMT_AMT_AMT,
+              String.format("%.2f", lineItem.coinsurancePaidByPayer.add(lineItem.paidByPayer)));
+          fieldValues.put(HOSPICE.REV_CNTR_TOT_CHRG_AMT,
+              String.format("%.2f", lineItem.cost));
+          fieldValues.put(HOSPICE.REV_CNTR_NCVRD_CHRG_AMT,
+              String.format("%.2f", lineItem.copayPaidByPatient
+              .add(lineItem.deductiblePaidByPatient).add(lineItem.patientOutOfPocket)));
+          if (lineItem.patientOutOfPocket.compareTo(Claim.ZERO_CENTS) == 0
+                  && lineItem.deductiblePaidByPatient.compareTo(Claim.ZERO_CENTS) == 0) {
+            // Not subject to deductible or coinsurance
+            fieldValues.put(HOSPICE.REV_CNTR_DDCTBL_COINSRNC_CD, "3");
+          } else if (lineItem.patientOutOfPocket.compareTo(Claim.ZERO_CENTS) > 0
+                  && lineItem.deductiblePaidByPatient.compareTo(Claim.ZERO_CENTS) > 0) {
+            // Subject to deductible and coinsurance
+            fieldValues.put(HOSPICE.REV_CNTR_DDCTBL_COINSRNC_CD, "0");
+          } else if (lineItem.patientOutOfPocket.compareTo(Claim.ZERO_CENTS) == 0) {
+            // Not subject to deductible
+            fieldValues.put(HOSPICE.REV_CNTR_DDCTBL_COINSRNC_CD, "1");
+          } else {
+            // Not subject to coinsurance
+            fieldValues.put(HOSPICE.REV_CNTR_DDCTBL_COINSRNC_CD, "2");
+          }
+          rifWriters.writeValues(HOSPICE.class, fieldValues);
+        }
+
+        if (claimLine == 1) {
+          // If claimLine still equals 1, then no line items were successfully added.
+          // Add a single top-level entry.
+          fieldValues.put(HOSPICE.CLM_LINE_NUM, Integer.toString(claimLine));
+          fieldValues.put(HOSPICE.REV_CNTR_DT, ExporterBase.bb2DateFromTimestamp(encounter.start));
+          fieldValues.put(HOSPICE.HCPCS_CD, "S9126"); // hospice per diem
+          rifWriters.writeValues(HOSPICE.class, fieldValues);
+        }
+      }
+      claimCount++;
+    }
+    return claimCount;
+  }
+
+
+  /**
+   * Thread safe singleton pattern adopted from
+   * https://stackoverflow.com/questions/7048198/thread-safe-singletons-in-java
+   */
+  private static class SingletonHolder {
+    /**
+     * Singleton instance of the BB2RIFExporter.
+     */
+    private static final BB2RIFExporter instance = new BB2RIFExporter();
+  }
+
+  /**
+   * Get the current instance of the BBExporter.
+   *
+   * @return the current instance of the BBExporter.
+   */
+  public static BB2RIFExporter getInstance() {
+    return SingletonHolder.instance;
+  }
+}
