+++ src/main/java/org/mitre/synthea/export/rif/ConsolidatedClaimLines.java
@@ -1,0 +1,91 @@
+package org.mitre.synthea.export.rif;
+
+import java.util.Collection;
+import java.util.Map;
+import java.util.TreeMap;
+import org.mitre.synthea.world.agents.Clinician;
+import org.mitre.synthea.world.concepts.Claim;
+import org.mitre.synthea.world.concepts.HealthRecord.Encounter;
+
+/**
+ * Utility class for consolidating multiple claim lines of the same type for the same clinician.
+ */
+class ConsolidatedClaimLines extends Claim.ClaimCost {
+
+  static class ConsolidatedClaimLine extends Claim.ClaimCost {
+
+    private int count;
+    private final String code;
+    private final String revCntr;
+    private final Clinician clinician;
+    private final long start;
+
+    public ConsolidatedClaimLine(Claim.ClaimCost initial, String code, String revCntr,
+            Encounter encounter) {
+      super(initial);
+      this.code = code;
+      this.revCntr = revCntr;
+      this.clinician = encounter.clinician;
+      this.start = encounter.start;
+      count = 1;
+    }
+
+    @Override
+    public void addCosts(Claim.ClaimCost other) {
+      super.addCosts(other);
+      count++;
+    }
+
+    public int getCount() {
+      return count;
+    }
+
+    public String getCode() {
+      return code;
+    }
+
+    public String getRevCntr() {
+      return revCntr;
+    }
+
+    public Clinician getClinician() {
+      return clinician;
+    }
+
+    public long getStart() {
+      return start;
+    }
+  }
+
+  private Map<String, ConsolidatedClaimLine> uniqueLineItems;
+
+  public ConsolidatedClaimLines() {
+    // use a sorted map to ensure we always emit claim lines in the same order
+    uniqueLineItems = new TreeMap<>();
+  } // use a sorted map to ensure we always emit claim lines in the same order
+
+  public void addClaimLine(String hcpcsCode, String revCntr, Claim.ClaimCost cost,
+          Encounter encounter) {
+    if (hcpcsCode == null) {
+      hcpcsCode = ""; // TreeMap doesn't like null keys unless you provide a custom comparator
+    }
+    if (revCntr == null) {
+      revCntr = "";
+    }
+    String clinicianId = "";
+    if (encounter.clinician != null && encounter.clinician.npi != null) {
+      clinicianId = encounter.clinician.npi;
+    }
+    String key = clinicianId + '-' + hcpcsCode + '-' + revCntr;
+    if (uniqueLineItems.containsKey(key)) {
+      uniqueLineItems.get(key).addCosts(cost);
+    } else {
+      uniqueLineItems.put(key, new ConsolidatedClaimLine(cost, hcpcsCode, revCntr, encounter));
+    }
+    this.addCosts(cost);
+  }
+
+  public Collection<ConsolidatedClaimLine> getLines() {
+    return uniqueLineItems.values();
+  }
+}
