+++ src/test/java/org/mitre/synthea/export/rif/BB2RIFMinimizerTest.java
@@ -1,4 +1,4 @@
-package org.mitre.synthea.export;
+package org.mitre.synthea.export.rif;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
